### 2019-01-22 #1195 @whedon
[PRE REVIEW]: pyclustering
**Submitting author:** @annoviko (<a href="http://orcid.org/0000-0002-9666-9141">Andrei Novikov</a>)
**Repository:** <a href="https://github.com/annoviko/pyclustering" target ="_blank">https://github.com/annoviko/pyclustering</a>
**Version:** 0.8.2
