### 2019-01-23 #1195 @annoviko
Hello, @arfon,
I have checked the list and found out that **adavidzh**, **dvysardana**, **pragyansmita** are suitable persons for review in line with the description (C/C++, Python, Cluster Analysis, Data Mining).
