### 2019-01-23 #1195 @arfon
> I have checked the list and found out that adavidzh, dvysardana, pragyansmita are suitable persons for review in line with the description (C/C++, Python, Cluster Analysis, Data Mining).

Thanks @annoviko. Before we decide upon reviewers, we need to have an editor assigned - this shouldn't take more than a week.

> The article devoted to 0.8.2 version that is on stable release branch '0.8.2.rel' (link: https://github.com/annoviko/pyclustering/tree/0.8.2.rel). DOI is already assigned for this release (link: https://zenodo.org/record/1491324#.XEgztVUzaUk) and it can be used by JOSS if it does not contradict the rules. Master branch already contains changes related to the new version of the library.

Is the paper (https://github.com/openjournals/joss-reviews/issues/1195#issuecomment-456493099) above not correct then? We can ask @whedon to compile a paper from a different branch in your repository.

Also, regarding the DOI & archive you already have - this might be OK, if your code is modified because of the review here we would ask you to make a new archive before publication.
