### 2019-01-23 828aeca
#381: Create skeleton of CLIQUE algorithm.
