+++ docs/citation.bib
@@ -579,6 +579,20 @@
     url             = {http://dx.doi.org/10.1016/0377-0427(87)90125-7},
     doi             = {10.1016/0377-0427(87)90125-7},
     publisher       = {Elsevier Science Publishers B. V.},
-} 
+}
 
 
+@article{article::clique::1,
+    author          = "Agrawal, Rakeshand Gehrke, Johannes and Gunopulos, Dimitrios and Raghavan, Prabhakar",
+    title           = "Automatic Subspace Clustering of High Dimensional Data",
+    journal         = "Data Mining and Knowledge Discovery",
+    year            = "2005",
+    month           = "Jul",
+    day             = "01",
+    volume          = "11",
+    number          = "1",
+    pages           = "5--33",
+    issn            = "1573-756X",
+    doi             = "10.1007/s10618-005-1396-1",
+    url             = "https://doi.org/10.1007/s10618-005-1396-1"
+}
