+++ pyclustering/cluster/clique.py
@@ -26,8 +26,175 @@
 
 
 class spatial_block:
+    """!
+    @brief Geometrical description of CLIQUE block in data space.
+    @details Provides services related to spatial functionality.
+
+    @see bang_block
+
+    """
+
+    def __init__(self, max_corner, min_corner):
+        """!
+        @brief Creates spatial block in data space.
+
+        @param[in] max_corner (array_like): Maximum corner coordinates of the block.
+        @param[in] min_corner (array_like): Minimal corner coordinates of the block.
+
+        """
+        self.__max_corner = max_corner
+        self.__min_corner = min_corner
+
+
+    def __str__(self):
+        """!
+        @brief Returns string block description.
+
+        @return String representation of the block.
+
+        """
+        return "(max: %s; min: %s)" % (self.__max_corner, self.__min_corner)
+
+
+    def __contains__(self, point):
+        """!
+        @brief Point is considered as contained if it lies in block (belong to it).
+
+        @return (bool) True if point is in block, otherwise False.
+
+        """
+        for i in range(len(point)):
+            if point[i] < self.__min_corner[i] or point[i] > self.__max_corner[i]:
+                return False
+
+        return True
+
+
+    def get_corners(self):
+        """!
+        @brief Return spatial description of current block.
+
+        @return (tuple) Pair of maximum and minimum corners (max_corner, min_corner).
+
+        """
+        return self.__max_corner, self.__min_corner
+
+
+    def is_neighbor(self, block):
+        """!
+        @brief Performs calculation to identify whether specified block is neighbor of current block.
+
+        @param[in] block (spatial_block): Another block that is check whether it is neighbor.
+
+        @return (bool) True is blocks are neighbors, False otherwise.
+
+        """
+        if block is not self:
+            block_max_corner, _ = block.get_corners()
+            dimension = len(block_max_corner)
+            neighborhood_score = self.__calculate_neighborhood(block_max_corner)
+
+            if neighborhood_score == dimension:
+                return True
+
+        return False
+
+
+    def __calculate_neighborhood(self, block_max_corner):
+        """!
+        @brief Calculates neighborhood score that defined whether blocks are neighbors.
+
+        @param[in] block_max_corner (list): Maximum coordinates of other block.
+
+        @return (uint) Neighborhood score.
+
+        """
+        dimension = len(block_max_corner)
+
+        length_edges = [self.__max_corner[i] - self.__min_corner[i] for i in range(dimension)]
+
+        neighborhood_score = 0
+        for i in range(dimension):
+            diff = abs(block_max_corner[i] - self.__max_corner[i])
+
+            if diff <= length_edges[i] + length_edges[i] * 0.0001:
+                neighborhood_score += 1
+
+        return neighborhood_score
+
+
+
+class clique_block:
     def __init__(self):
+        self.__logical_location = []
+        self.__spatial_location = None
+        self.__points = []
+
+    @property
+    def logical_location(self):
+        return self.__logical_location
+
+    @logical_location.setter
+    def logical_location(self, location):
+        self.__logical_location = location
+
+    @property
+    def spatial_location(self):
+        return self.__spatial_location
+
+    @spatial_location.setter
+    def spatial_location(self, location):
+        self.__spatial_location = location
+
+    @property
+    def points(self):
+        return self.__points
+
+    def capture_points(self, data, point_availability):
+        for index_point in range(len(data)):
+            if (point_availability[index_point] is True) and (data[index_point] in self.__spatial_location):
+                self.__points.append(index_point)
+                point_availability[index_point] = False
+
+
[…]
