### 2019-01-28 #1195 @danielskatz
👋 @arfon - is this ready to go to an editor in your opinion?  If so, can you suggest one?  If not, what else do you think is needed?
