### 2019-01-29 12dfc1c
#381: CLIQUE and examples for it.
