+++ pyclustering/cluster/clique.py
@@ -80,55 +80,13 @@ class spatial_block:
         return self.__max_corner, self.__min_corner
 
 
-    def is_neighbor(self, block):
-        """!
-        @brief Performs calculation to identify whether specified block is neighbor of current block.
-
-        @param[in] block (spatial_block): Another block that is check whether it is neighbor.
-
-        @return (bool) True is blocks are neighbors, False otherwise.
-
-        """
-        if block is not self:
-            block_max_corner, _ = block.get_corners()
-            dimension = len(block_max_corner)
-            neighborhood_score = self.__calculate_neighborhood(block_max_corner)
-
-            if neighborhood_score == dimension:
-                return True
-
-        return False
-
-
-    def __calculate_neighborhood(self, block_max_corner):
-        """!
-        @brief Calculates neighborhood score that defined whether blocks are neighbors.
-
-        @param[in] block_max_corner (list): Maximum coordinates of other block.
-
-        @return (uint) Neighborhood score.
-
-        """
-        dimension = len(block_max_corner)
-
-        length_edges = [self.__max_corner[i] - self.__min_corner[i] for i in range(dimension)]
-
-        neighborhood_score = 0
-        for i in range(dimension):
-            diff = abs(block_max_corner[i] - self.__max_corner[i])
-
-            if diff <= length_edges[i] + length_edges[i] * 0.0001:
-                neighborhood_score += 1
-
-        return neighborhood_score
-
-
 
 class clique_block:
     def __init__(self):
         self.__logical_location = []
         self.__spatial_location = None
         self.__points = []
+        self.__belong = False
 
     @property
     def logical_location(self):
@@ -150,6 +108,15 @@ class clique_block:
     def points(self):
         return self.__points
 
+    @property
+    def belong(self):
+        return self.__belong
+
+    @belong.setter
+    def belong(self, value):
+        self.__belong = value
+
+
     def capture_points(self, data, point_availability):
         for index_point in range(len(data)):
             if (point_availability[index_point] is True) and (data[index_point] in self.__spatial_location):
@@ -157,20 +124,21 @@ class clique_block:
                 point_availability[index_point] = False
 
 
-    def is_neighbor(self, block):
+    def get_location_neighbors(self, edge):
+        neighbors = []
+
         for index_dimension in range(len(self.__logical_location)):
-            if not self.__is_dimension_neighbor(self.__logical_location[index_dimension], block.logical_location[index_dimension]):
-                return False
+            if self.__logical_location[index_dimension] + 1 < edge:
+                position = self.__logical_location[:]
+                position[index_dimension] += 1
+                neighbors.append(position)
 
-        return True
+            if self.__logical_location[index_dimension] - 1 >= 0:
+                position = self.__logical_location[:]
+                position[index_dimension] -= 1
+                neighbors.append(position)
 
-
-    def get_neighbors(self):
-        pass
-
-
-    def __is_dimension_neighbor(self, location1, location2):
-        return location1 + 1 == location2 or location1 - 1 == location2
+        return neighbors
 
 
 
@@ -206,12 +174,17 @@ class clique:
         self.__clusters = []
         self.__noise = []
 
+        self.__cells = []
+        self.__cells_map = {}
+
         self.__validate_arguments()
 
 
     def process(self):
-        cells = self.__create_grid()
-        self.__allocate_clusters(cells)
+        self.__create_grid()
+        self.__allocate_clusters()
+
+        self.__cells_map.clear()
 
 
     def get_clusters(self):
@@ -233,23 +206,48 @@ class clique:
[…]
