### 2019-01-29 36d3a7c
#381: CLIQUE and examples for it (investigation for LSUN is required).
