### 2019-01-29 af2a672
#381: CLIQUE implemetation (updates and corrections), CLIQUE visualizer implementation. Test template is prepared.
