+++ pyclustering/cluster/tests/unit/ut_clique.py
@@ -1,0 +1,34 @@
+"""!
+
+@brief Unit-tests for CLIQUE algorithm.
+
+@authors Andrei Novikov (pyclustering@yandex.ru)
+@date 2014-2019
+@copyright GNU Public License
+
+@cond GNU_PUBLIC_LICENSE
+    PyClustering is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    PyClustering is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+@endcond
+
+"""
+
+import unittest
+
+# Generate images without having a window appear.
+import matplotlib
+matplotlib.use('Agg')
+
+from pyclustering.cluster.tests.clique_templates import clique_test_template
+
+from pyclustering.samples.definitions import SIMPLE_SAMPLES
