### 2019-01-30 7d829f4
#381: CLIQUE python implementation (update examples and add unit-tests).
