+++ pyclustering/cluster/tests/unit/ut_clique.py
@@ -29,6 +29,148 @@ import unittest
 import matplotlib
 matplotlib.use('Agg')
 
+from pyclustering.cluster.clique import clique_block
 from pyclustering.cluster.tests.clique_templates import clique_test_template
 
-from pyclustering.samples.definitions import SIMPLE_SAMPLES
+from pyclustering.tests.assertion import assertion
+
+from pyclustering.samples.definitions import SIMPLE_SAMPLES, FCPS_SAMPLES
+
+
+class clique_unit_test(unittest.TestCase):
+    def test_clustering_sample_simple_1(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 8, 0, [5, 5], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 7, 0, [5, 5], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 6, 0, [5, 5], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 5, 0, [5, 5], 0, False)
+
+    def test_clustering_sample_simple_1_one_cluster(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 1, 0, [10], 0, False)
+
+    def test_clustering_diagonal_blocks_arent_neoghbors(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 2, 0, [5, 5], 0, False)
+
+    def test_clustering_sample_simple_1_noise_only(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 6, 1000, [], 10, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 6, 10, [], 10, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 2, 5, [], 10, False)
+
+    def test_clustering_sample_simple_2(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE2, 7, 0, [5, 8, 10], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE2, 6, 0, [5, 8, 10], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE2, 1, 0, [23], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE2, 6, 500, [], 23, False)
+
+    def test_clustering_sample_simple_3(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE3, 9, 0, [10, 10, 10, 30], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE3, 8, 0, [10, 10, 10, 30], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE3, 1, 0, [60], 0, False)
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE3, 6, 500, [], 60, False)
+
+    def test_clustering_sample_simple_3_one_point_noise(self):
+        clique_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE3, 2, 9, [59], 1, False)
+
[…]
