+++ pyclustering/cluster/tests/unit/ut_bang.py
@@ -34,7 +34,7 @@ from pyclustering.cluster.tests.bang_templates import bang_test_template
 from pyclustering.samples.definitions import SIMPLE_SAMPLES
 
 
-class bsas_unit_test(unittest.TestCase):
+class bang_unit_test(unittest.TestCase):
     def test_clustering_sample_simple_1(self):
         bang_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 8, 0.0, [5, 5], 0, False)
         bang_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 7, 0.0, [5, 5], 0, False)
@@ -43,6 +43,9 @@ class bsas_unit_test(unittest.TestCase):
 
     def test_clustering_sample_simple_1_one_cluster(self):
         bang_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 1, 0.0, [10], 0, False)
+
+    def test_clustering_diagonal_neighbors(self):
+        bang_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 3, 0.0, [10], 0, False)
 
     def test_clustering_sample_simple_1_noise_only(self):
         bang_test_template.clustering(SIMPLE_SAMPLES.SAMPLE_SIMPLE1, 6, 1000.0, [], 10, False)
