### 2019-01-31 ae77281
#381: Expand cluster optimization.
