### 2019-01-31 bd87170
#381: C++ implementation of CLIQUE (compilable skeleton).
