### 2019-02-01 cdbc506
#381: Expand cluster optimization - version 2.
