### 2019-02-02 daa7482
#381: CLIQUE C++ final implementation, unit-tests for it.
