+++ ccore/tst/utest-clique.cpp
@@ -1,0 +1,183 @@
+/**
+*
+* @authors Andrei Novikov (pyclustering@yandex.ru)
+* @date 2014-2019
+* @copyright GNU Public License
+*
+* GNU_PUBLIC_LICENSE
+*   pyclustering is free software: you can redistribute it and/or modify
+*   it under the terms of the GNU General Public License as published by
+*   the Free Software Foundation, either version 3 of the License, or
+*   (at your option) any later version.
+*
+*   pyclustering is distributed in the hope that it will be useful,
+*   but WITHOUT ANY WARRANTY; without even the implied warranty of
+*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+*   GNU General Public License for more details.
+*
+*   You should have received a copy of the GNU General Public License
+*   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*
+*/
+
+
+#include "gtest/gtest.h"
+
+#include "cluster/clique.hpp"
+
+#include "samples.hpp"
+
+#include "utenv_check.hpp"
+
+
+using namespace ccore::clst;
+
+
+static void
+template_clique_length_process_data(
+    const dataset_ptr p_data,
+    const std::size_t p_intervals,
+    const std::size_t p_threshold,
+    const std::vector<std::size_t> & p_expected_cluster_length,
+    const std::size_t p_expected_noise_length) 
+{
+    clique_data output_result;
+    clique solver(p_intervals, p_threshold);
+    solver.process(*p_data, output_result);
+
+    const dataset & data = *p_data;
+    const cluster_sequence & actual_clusters = output_result.clusters();
+    const noise & noise = output_result.noise();
+    const clique_block_sequence & blocks = output_result.blocks();
+
+    ASSERT_CLUSTER_NOISE_SIZES(data, actual_clusters, p_expected_cluster_length, noise, p_expected_noise_length);
+}
+
+TEST(utest_clique, allocation_sample_simple_01) {
+    const std::vector<std::size_t> expected_clusters_length = { 5, 5 };
+    template_clique_length_process_data(simple_sample_factory::create_sample(SAMPLE_SIMPLE::SAMPLE_SIMPLE_01), 8, 0, expected_clusters_length, 0);
+    template_clique_length_process_data(simple_sample_factory::create_sample(SAMPLE_SIMPLE::SAMPLE_SIMPLE_01), 7, 0, expected_clusters_length, 0);
+    template_clique_length_process_data(simple_sample_factory::create_sample(SAMPLE_SIMPLE::SAMPLE_SIMPLE_01), 6, 0, expected_clusters_length, 0);
+    template_clique_length_process_data(simple_sample_factory::create_sample(SAMPLE_SIMPLE::SAMPLE_SIMPLE_01), 5, 0, expected_clusters_length, 0);
+}
+
+TEST(utest_clique, allocation_sample_simple_01_one_cluster) {
+    const std::vector<std::size_t> expected_clusters_length = { 10 };
+    template_clique_length_process_data(simple_sample_factory::create_sample(SAMPLE_SIMPLE::SAMPLE_SIMPLE_01), 1, 0, expected_clusters_length, 0);
+}
+
+TEST(utest_clique, allocation_sample_simple_01_only_noise) {
[…]
