### 2019-02-02 1bb32c0
#381: ASSERT_CLUSTER_NOISE_SIZES correction and additional checking for CLIQUE's blocks.
