+++ ccore/tst/utenv_check.cpp
@@ -37,7 +37,7 @@ void ASSERT_CLUSTER_SIZES(
     const std::vector<size_t> & p_expected_cluster_length,
     const index_sequence & p_indexes)
 {
-    ASSERT_CLUSTER_NOISE_SIZES(p_data, p_actual_clusters, p_expected_cluster_length, {}, -1, p_indexes);
+    ASSERT_CLUSTER_NOISE_SIZES(p_data, p_actual_clusters, p_expected_cluster_length, { }, -1, p_indexes);
 }
 
 
@@ -75,7 +75,10 @@ void ASSERT_CLUSTER_NOISE_SIZES(
     ASSERT_EQ(total_size, unique_objects.size());
 
     if (!p_expected_cluster_length.empty()) {
-        std::size_t expected_total_size = std::accumulate(p_expected_cluster_length.cbegin(), p_expected_cluster_length.cend(), (std::size_t) 0) + p_expected_noise_length;
+        std::size_t expected_total_size = std::accumulate(p_expected_cluster_length.cbegin(), p_expected_cluster_length.cend(), (std::size_t) 0);
+        if (p_expected_noise_length != (std::size_t) -1) {
+             expected_total_size += p_expected_noise_length;
+        }
 
         ASSERT_EQ(expected_total_size, total_size);
 
