### 2019-02-05 8a50e94
[pyclustering.utils] Remove not used functions.
