### 2019-02-06 b7f7018
#381: CLIQUE documentation, examples and meta-information.
