+++ pyclustering/cluster/clique.py
@@ -27,6 +27,7 @@
 
 import itertools
 
+from pyclustering.cluster import cluster_visualizer
 from pyclustering.core.wrapper import ccore_library
 
 import pyclustering.core.clique_wrapper as wrapper
@@ -45,10 +46,26 @@ except Exception as error_instance:
 
 
 class clique_visualizer:
+    """!
+    @brief Visualizer of CLIQUE algorithm's results.
+    @details CLIQUE visualizer provides visualization services that are specific for CLIQUE algorithm, for example,
+              to display grid and its density.
+
+    """
+
     __maximum_density_alpha = 0.6
 
     @staticmethod
     def show_grid(cells, data):
+        """!
+        @brief Show CLIQUE blocks as a grid in data space.
+        @details Each block contains points and according to this density is displayed. CLIQUE grid helps to visualize
+                  grid that was used for clustering process.
+
+        @param[in] cells (list): List of cells that is produced by CLIQUE algorithm.
+        @param[in] data (array_like): Input data that was used for clustering process.
+
+        """
         dimension = cells[0].dimensions
 
         amount_canvases = 1
@@ -67,6 +84,22 @@ class clique_visualizer:
             clique_visualizer.__draw_two_dimension_data(ax, data, pairs[index])
 
         plt.show()
+
+
+    @staticmethod
+    def show_clusters(data, clusters, noise=None):
+        """!
+        @brief Display CLIQUE clustering results.
+
+        @param[in] data (list): Data that was used for clustering.
+        @param[in] clusters (array_like): Clusters that were allocated by the algorithm.
+        @param[in] noise (array_like): Noise that were allocated by the algorithm.
+
+        """
+        visualizer = cluster_visualizer()
+        visualizer.append_clusters(clusters, data)
+        visualizer.append_cluster(noise or [], data, marker='x')
+        visualizer.show()
 
 
     @staticmethod
@@ -191,52 +224,138 @@ class spatial_block:
 
 
 class clique_block:
+    """!
+    @brief CLIQUE block contains information about its logical location in grid, spatial location in data space and
+            points that are covered by the block.
+
+    """
+
     def __init__(self, logical_location=None, spatial_location=None, points=None, visited=False):
+        """!
+        @brief Initializes CLIQUE block.
+
+        @param[in] logical_location (list): Logical location of the block in CLIQUE grid.
+        @param[in] spatial_location (spatial_block): Spatial location in data space.
+        @param[in] points (array_like): Points that belong to this block (can be obtained by method 'capture_points',
+                    this parameter is used by CLIQUE in case of processing by C++ implementation when clustering
+                    result are passed back to Python code.
+        @param[in] visited (bool): Marks if block is visited during clustering process.
+
+        """
         self.__logical_location = logical_location or []
         self.__spatial_location = spatial_location
         self.__points = points or []
         self.__visited = visited
 
     def __str__(self):
+        """!
+        @brief Returns string representation of the block using its logical location in CLIQUE grid.
+
+        """
         return str(self.__logical_location)
 
     def __repr__(self):
+        """!
+        @brief Returns string representation of the block using its logical location in CLIQUE grid.
+
+        """
         return str(self.__logical_location)
 
     @property
     def logical_location(self):
+        """!
+        @brief Logical location is represented by coordinates in CLIQUE grid, for example, in case of 2x2 grid blocks
+                may have following coordinates: [0, 0], [0, 1], [1, 0], [1, 1].
+        @return (list) Logical location of the block in CLIQUE grid.
+
+        """
         return self.__logical_location
 
     @logical_location.setter
     def logical_location(self, location):
+        """!
+        @brief Assign logical location to CLIQUE block.
+
[…]
