### 2019-02-06 506deb6
#381: CLIQUE documentation.
