### 2019-02-12 #1230 @adavidzh
Hi @poulson, I'm not sure how to check the "release version".

The repo has a [`0.8.2.rel` branch whose `README.rst` identifies the version as `0.8.2`](https://github.com/annoviko/pyclustering/tree/0.8.2.rel). [PyPi mentions `0.8.2`](https://pypi.org/project/pyclustering/). So, while things looks consistent, I'm not sure if "the release version given match[es] the GitHub release (0.8.2)".
