### 2019-02-20 #1230 @poulson
@adavidzh I apologize for missing the ping on this! I will reach out to the author to confirm.

@annoviko Andrei, Could you confirm that `0.8.2rel` is the proper branch? It would be great if you could take a minute and create a `0.8.2` branch so that there is a complete match.
