### 2019-02-20 #1230 @annoviko
Hello, @poulson, yes I can confirm that '0.8.2.rel' is the proper branch that was created to release version 0.8.2. There is already '0.8.2' tag as you can see it on pyclustering release page: https://github.com/annoviko/pyclustering/tree/0.8.2 (release page: https://github.com/annoviko/pyclustering/releases/tag/0.8.2).
