### 2019-02-21 #1230 @annoviko
@adavidzh, you welcome, do not hesitate to ask me any questions related to pyclustering.
