### 2019-03-04 annoviko/pyclustering#492 @adavidzh
Installation instructions
Somehow [this line in `README.rst` is broken](https://github.com/annoviko/pyclustering/blame/506deb69fcb3e66b61c938055886f44f35283c93/README.rst#L73). It should probably be `cd ccore`.
