### 2019-03-04 #1230 @adavidzh
I have run into the following issues:
- manual installation instructions need to be updated: <https://github.com/annoviko/pyclustering/issues/492>.
- `pip` installation works, _even_ in unsupported architectures; I'm on `Darwin` so I'll have to fire-up a VM to continue.

The paper itself could benefit a lot from:
- putting in tables the running text where available algorithms are listed.
- breaking the monolithic `Summary` section into the four main parts of the library, viz. `Clustering algorithms`, `Oscillatory networks and neural networks`, `Graph Coloring Algorithms`, and `Containers`, each with a table and its own statement of need.
