### 2019-03-04 276c7ff
#492: Install instructions for 0.8.2.rel.
