### 2019-03-04 annoviko/pyclustering#492 @annoviko
@adavidzh , instructions are updated for 0.8.2.rel and for master branch.
