### 2019-03-04 #1230 @annoviko
@adavidzh , I have updated instruction in line with annoviko/pyclustering#492.

About installation on MAC OS using `pip`. Yes, it works, but `CCORE `(fast code - C/C++ implementation) is not supported, python implementation is used instead. `CCORE `is build only for Linux and Windows as you can see it in CI:
- Travis-CI (Linux builds): https://travis-ci.org/annoviko/pyclustering
- AppVeyor (Windows builds): https://ci.appveyor.com/project/annoviko/pyclustering/branch/master

But pyclustering itself works on every other platform if it is supported by python and libraries that are used by pyclustering.

About the paper. I will provide changes soon.
