### 2019-03-04 annoviko/pyclustering#492 @adavidzh
Thanks.
