### 2019-03-04 annoviko/pyclustering#493 @adavidzh
Example in README
[One example in `README.rst`](https://github.com/annoviko/pyclustering/blame/a201fb7ca11a3d5637984e4d157e6b50ef3d9234/README.rst#L275) seems to depend on an undefined argument.

This should be rewritten to work out of the box.
