### 2019-03-04 #1230 @adavidzh
Thanks @annoviko. I now understand that only the `CCORE` does not work in `Darwin`. I am working through the examples and reported one issue here https://github.com/annoviko/pyclustering/issues/493. As for the paper, that would be great, thanks.
