### 2019-03-04 annoviko/pyclustering#493 @adavidzh
Also the last example does not work out of the box because of [`SIMPLE_SAMPLES.SAMPLE_SIMPLE1`](https://github.com/annoviko/pyclustering/blame/a201fb7ca11a3d5637984e4d157e6b50ef3d9234/README.rst#L297) not being defined.
