+++ docs/citation.bib
@@ -596,3 +596,13 @@
     doi             = "10.1007/s10618-005-1396-1",
     url             = "https://doi.org/10.1007/s10618-005-1396-1"
 }
+
+
+@book{book::pattern_recognition_with_fuzzy,
+    author          = {Bezdek, James C.},
+    title           = {Pattern Recognition with Fuzzy Objective Function Algorithms},
+    year            = {1981},
+    isbn            = {0306406713},
+    publisher       = {Kluwer Academic Publishers},
+    address         = {Norwell, MA, USA},
+}
