+++ README.rst
@@ -120,55 +120,110 @@ Brief Overview of the Library Content
 
 **Clustering algorithms and methods (module pyclustering.cluster):**
 
-- Agglomerative [Python, C++]
-- BANG [Python]
-- BIRCH [Python]
-- BSAS [Python, C++]
-- CLARANS [Python]
-- CURE [Python, C++]
-- DBSCAN [Python, C++]
-- Elbow [Python, C++]
-- EMA [Python]
-- GA (Genetic Algorithm) [Python, C++]
-- HSyncNet [Python, C++]
-- K-Means [Python, C++]
-- K-Means++ [Python, C++]
-- K-Medians [Python, C++]
-- K-Medoids (PAM) [Python, C++]
-- MBSAS [Python, C++]
-- OPTICS [Python, C++]
-- ROCK [Python, C++]
-- Silhouette [Python]
-- SOM-SC [Python, C++]
-- SyncNet [Python, C++]
-- Sync-SOM [Python]
-- TTSAS [Python, C++]
-- X-Means [Python, C++]
++------------------------+---------+-----+
+| Algorithm              | Python  | C++ |
++========================+=========+=====+
+| Agglomerative          | ✓       | ✓   |
++------------------------+---------+-----+
+| BANG                   | ✓       |     |
++------------------------+---------+-----+
+| BIRCH                  | ✓       |     |
++------------------------+---------+-----+
+| BSAS                   | ✓       | ✓   |
++------------------------+---------+-----+
+| CLARANS                | ✓       |     |
++------------------------+---------+-----+
+| CLIQUE                 | ✓       | ✓   |
++------------------------+---------+-----+
+| CURE                   | ✓       | ✓   |
++------------------------+---------+-----+
+| DBSCAN                 | ✓       | ✓   |
++------------------------+---------+-----+
+| Elbow                  | ✓       | ✓   |
++------------------------+---------+-----+
+| EMA                    | ✓       |     |
++------------------------+---------+-----+
+| GA (Genetic Algorithm) | ✓       | ✓   |
++------------------------+---------+-----+
+| HSyncNet               | ✓       | ✓   |
++------------------------+---------+-----+
+| K-Means                | ✓       | ✓   |
++------------------------+---------+-----+
+| K-Means++              | ✓       | ✓   |
++------------------------+---------+-----+
+| K-Medians              | ✓       | ✓   |
++------------------------+---------+-----+
+| K-Medoids (PAM)        | ✓       | ✓   |
++------------------------+---------+-----+
+| MBSAS                  | ✓       | ✓   |
++------------------------+---------+-----+
+| OPTICS                 | ✓       | ✓   |
++------------------------+---------+-----+
+| ROCK                   | ✓       | ✓   |
++------------------------+---------+-----+
+| Silhouette             | ✓       |     |
++------------------------+---------+-----+
+| SOM-SC                 | ✓       | ✓   |
++------------------------+---------+-----+
+| SyncNet                | ✓       | ✓   |
++------------------------+---------+-----+
+| Sync-SOM               | ✓       |     |
++------------------------+---------+-----+
+| TTSAS                  | ✓       | ✓   |
++------------------------+---------+-----+
+| X-Means                | ✓       | ✓   |
++------------------------+---------+-----+
 
 
 **Oscillatory networks and neural networks (module pyclustering.nnet):**
 
-- CNN (Chaotic Neural Network) [Python] 
-- fSync (Oscillatory network based on Landau-Stuart equation and Kuramoto model) [Python] 
-- HHN (Oscillatory network based on Hodgkin-Huxley model) [Python, C++]
-- Hysteresis Oscillatory Network [Python]
-- LEGION (Local Excitatory Global Inhibitory Oscillatory Network) [Python, C++]
[…]
