### 2019-03-09 annoviko/pyclustering#493 @annoviko
Hello, @adavidzh,
README is updated: examples are rechecked, updated, and several new examples are added.
