### 2019-03-09 9871b1d
[.] Update paper in line with comments from reviewer @adavidzh.
