+++ paper/paper.md
@@ -35,12 +35,118 @@ The Python implementation uses SciPy [@SciPy], MatPlotLib [@Hunter2007], NumPy a
 
 One of the unique features of the library is a collection of oscillatory networks for cluster analysis, graph coloring, and image segmentation. Oscillatory networks are biologically plausible neural networks that use synchronization processes for solving practical problems. Formally, oscillatory neural networks are nonlinear dynamic systems in which neuron is an oscillating element called an oscillator. There is an assumption that the synchronization processes between neurons in the brain are used to implement cognitive functions [@Novikov2014][@Cumin2007]. Thus, oscillatory networks are of great interest because they allow to research mechanisms that synchronize the neuronal activity at the model level.
 
-PyClustering library includes following clustering algorithms: agglomerative, BANG, BIRCH, BSAS, CLARANS, CURE, CURE, DBSCAN, Elbow, expectation-maximization algorithm, genetic algorithm, hierarchical SyncNet, K-Means, K-Means++, K-Medians, K-Medoids (PAM), MBSAS, OPTICS, ROCK, Silhouette, SOM-SC, SyncNet, Sync-SOM, TTSAS, X-Means. These algorithms and methods are located in module 'pyclustering.cluster' in case of Python and in namespace 'ccore.clst' in case of C++.
-
-Following oscillatory and neural networks are implemented in PyClustering: CNN (chaotic neural network), fSync (oscillatory network based on Landau-Stuart equation and Kuramoto model), HHN (oscillatory network based on Hodgkin-Huxley model), hysteresis oscillatory network, LEGION (local excitatory global inhibitory oscillatory network), PCNN (pulse-coupled neural network), SOM (self-organized map), Sync (oscillatory network based on Kuramoto model), SyncPR (oscillatory network based on Kuramoto model for pattern recognition), SyncSegm (oscillatory network based on Kuramoto model for image segmentation).
-
-Additionally, the library contains several graph coloring algorithms such as DSatur, Hysteresis, GColorSync.
-
 PyClustering library is available on PyPi and github repository. Since the first release on PyPi in 2014, it has been downloaded more than 141.000 times. Quality of the library is supported by static and dynamic analyzers such as cppcheck, scan-build, valgrid [@Nethercote2007], including compilers gcc, clang, VS2015. Code coverage is more than 93% that is ensured by unit and integration tests (total amount of tests is more than 2.200). Each commit to repository triggers building, analysis, and testing on CI services such as travis-ci, appveyor. PyClustering provides fully documented code for each library’s version including examples, math and algorithms description, installation instructions. The API documentation is generated by doxygen without any warnings and notes to ensure completeness.
+
+# Clustering Algorithms
+
+Algorithms and methods are located in module 'pyclustering.cluster' in case of Python and in namespace 'ccore.clst' in case of C++.
+
++------------------------+---------+-----+
+| Algorithm              | Python  | C++ |
++========================+=========+=====+
+| Agglomerative          | ✓       | ✓   |
++------------------------+---------+-----+
+| BANG                   | ✓       |     |
++------------------------+---------+-----+
+| BIRCH                  | ✓       |     |
++------------------------+---------+-----+
+| BSAS                   | ✓       | ✓   |
++------------------------+---------+-----+
+| CLARANS                | ✓       |     |
++------------------------+---------+-----+
+| CLIQUE                 | ✓       | ✓   |
++------------------------+---------+-----+
+| CURE                   | ✓       | ✓   |
++------------------------+---------+-----+
+| DBSCAN                 | ✓       | ✓   |
++------------------------+---------+-----+
+| Elbow                  | ✓       | ✓   |
++------------------------+---------+-----+
[…]
