### 2019-03-09 #1230 @annoviko
@adavidzh, I have updated the paper (see 'master' branch: `paper/paper.md`). And updated example section in `README` file ('master' and '0.8.2.rel' branches).
