### 2019-03-10 8d7095e
[.][no-build] Update paper, README and CI scripts to avoid building when it is not required.
