+++ paper/paper.md
@@ -44,55 +44,55 @@ Algorithms and methods are located in module 'pyclustering.cluster' in case of P
 +------------------------+---------+-----+
 | Algorithm              | Python  | C++ |
 +========================+=========+=====+
-| Agglomerative          | ✓       | ✓   |
+| Agglomerative          | v       | v   |
 +------------------------+---------+-----+
-| BANG                   | ✓       |     |
+| BANG                   | v       |     |
 +------------------------+---------+-----+
-| BIRCH                  | ✓       |     |
+| BIRCH                  | v       |     |
 +------------------------+---------+-----+
-| BSAS                   | ✓       | ✓   |
+| BSAS                   | v       | v   |
 +------------------------+---------+-----+
-| CLARANS                | ✓       |     |
+| CLARANS                | v       |     |
 +------------------------+---------+-----+
-| CLIQUE                 | ✓       | ✓   |
+| CLIQUE                 | v       | v   |
 +------------------------+---------+-----+
-| CURE                   | ✓       | ✓   |
+| CURE                   | v       | v   |
 +------------------------+---------+-----+
-| DBSCAN                 | ✓       | ✓   |
+| DBSCAN                 | v       | v   |
 +------------------------+---------+-----+
-| Elbow                  | ✓       | ✓   |
+| Elbow                  | v       | v   |
 +------------------------+---------+-----+
-| EMA                    | ✓       |     |
+| EMA                    | v       |     |
 +------------------------+---------+-----+
-| GA (Genetic Algorithm) | ✓       | ✓   |
+| GA (Genetic Algorithm) | v       | v   |
 +------------------------+---------+-----+
-| HSyncNet               | ✓       | ✓   |
+| HSyncNet               | v       | v   |
 +------------------------+---------+-----+
-| K-Means                | ✓       | ✓   |
+| K-Means                | v       | v   |
 +------------------------+---------+-----+
-| K-Means++              | ✓       | ✓   |
+| K-Means++              | v       | v   |
 +------------------------+---------+-----+
-| K-Medians              | ✓       | ✓   |
+| K-Medians              | v       | v   |
 +------------------------+---------+-----+
-| K-Medoids (PAM)        | ✓       | ✓   |
+| K-Medoids (PAM)        | v       | v   |
 +------------------------+---------+-----+
-| MBSAS                  | ✓       | ✓   |
+| MBSAS                  | v       | v   |
 +------------------------+---------+-----+
-| OPTICS                 | ✓       | ✓   |
+| OPTICS                 | v       | v   |
 +------------------------+---------+-----+
-| ROCK                   | ✓       | ✓   |
+| ROCK                   | v       | v   |
 +------------------------+---------+-----+
-| Silhouette             | ✓       |     |
+| Silhouette             | v       |     |
 +------------------------+---------+-----+
-| SOM-SC                 | ✓       | ✓   |
+| SOM-SC                 | v       | v   |
 +------------------------+---------+-----+
-| SyncNet                | ✓       | ✓   |
+| SyncNet                | v       | v   |
 +------------------------+---------+-----+
-| Sync-SOM               | ✓       |     |
+| Sync-SOM               | v       |     |
 +------------------------+---------+-----+
-| TTSAS                  | ✓       | ✓   |
+| TTSAS                  | v       | v   |
 +------------------------+---------+-----+
-| X-Means                | ✓       | ✓   |
+| X-Means                | v       | v   |
 +------------------------+---------+-----+
 
 # Oscillatory Networks and Neural Networks
@@ -102,25 +102,25 @@ Networks are located in module 'pyclustering.nnet' in case of Python and in name
 +--------------------------------------------------------------------------------+---------+-----+
 | Model                                                                          | Python  | C++ |
[…]
