### 2019-03-10 #1230 @annoviko
Just for your information: PDF generator does not support symbol '✓' (see @wheldon's previous post - tables have empty columns), therefore I have changed it to 'v'.
