### 2019-03-10 annoviko/pyclustering#493 @adavidzh
This is much better, thanks!

Right now there is one problem on the OPTICS example. It looks like there is a `optics_instance.process()` missing since at the end of it we're just getting nothing:
```
>>> list(map(lambda x:type(x), [clusters,noise]))
[<class 'NoneType'>, <class 'NoneType'>]
```

It would also be nice if there were a visualization for the result, which I would not know exactly how to add.
