### 2019-03-10 #1230 @adavidzh
- Paper: this looks much better, thanks. Based on the [instructions](https://joss.readthedocs.io/en/latest/submitting.html), I suggest trying [`$\checkmark$`, `$\Checkmark$`, `$\ding{51}$`, `$\ballotcheck$`, and others like those from `fontawesome`](http://tug.ctan.org/tex-archive/info/symbols/comprehensive/symbols-a4.pdf#page=134) in a single try and seeing if JOSS supports any of them.
- `README`: also good progress; annoviko/pyclustering#492 is not yet finished but should be close to that.

I am planning to make a PR against `paper.md` regarding some text issues _after_ you give a try to the  checkmark options above. If there is prior art somewhere on how to do it, I couldn't find it.
