### 2019-03-10 9cbe8ec
#493: Update OPTICS examples for README and for documentation itself.
