### 2019-03-10 bb79508
#493: [0.8.2.rel] Update OPTICS examples for README and for documentation itself.
