### 2019-03-10 f1c8e56
[.][no-build] Try to use latex checkmark in the paper.
