+++ paper/paper.md
@@ -41,112 +41,112 @@ PyClustering library is available on PyPi and github repository. Since the first
 
 Algorithms and methods are located in module 'pyclustering.cluster' in case of Python and in namespace 'ccore.clst' in case of C++.
 
-+------------------------+---------+-----+
-| Algorithm              | Python  | C++ |
-+========================+=========+=====+
-| Agglomerative          | v       | v   |
-+------------------------+---------+-----+
-| BANG                   | v       |     |
-+------------------------+---------+-----+
-| BIRCH                  | v       |     |
-+------------------------+---------+-----+
-| BSAS                   | v       | v   |
-+------------------------+---------+-----+
-| CLARANS                | v       |     |
-+------------------------+---------+-----+
-| CLIQUE                 | v       | v   |
-+------------------------+---------+-----+
-| CURE                   | v       | v   |
-+------------------------+---------+-----+
-| DBSCAN                 | v       | v   |
-+------------------------+---------+-----+
-| Elbow                  | v       | v   |
-+------------------------+---------+-----+
-| EMA                    | v       |     |
-+------------------------+---------+-----+
-| GA (Genetic Algorithm) | v       | v   |
-+------------------------+---------+-----+
-| HSyncNet               | v       | v   |
-+------------------------+---------+-----+
-| K-Means                | v       | v   |
-+------------------------+---------+-----+
-| K-Means++              | v       | v   |
-+------------------------+---------+-----+
-| K-Medians              | v       | v   |
-+------------------------+---------+-----+
-| K-Medoids (PAM)        | v       | v   |
-+------------------------+---------+-----+
-| MBSAS                  | v       | v   |
-+------------------------+---------+-----+
-| OPTICS                 | v       | v   |
-+------------------------+---------+-----+
-| ROCK                   | v       | v   |
-+------------------------+---------+-----+
-| Silhouette             | v       |     |
-+------------------------+---------+-----+
-| SOM-SC                 | v       | v   |
-+------------------------+---------+-----+
-| SyncNet                | v       | v   |
-+------------------------+---------+-----+
-| Sync-SOM               | v       |     |
-+------------------------+---------+-----+
-| TTSAS                  | v       | v   |
-+------------------------+---------+-----+
-| X-Means                | v       | v   |
-+------------------------+---------+-----+
++------------------------+--------------+--------------+
+| Algorithm              | Python       | C++          |
++========================+==============+==============+
+| Agglomerative          | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| BANG                   | $\checkmark$ |              |
++------------------------+--------------+--------------+
+| BIRCH                  | $\checkmark$ |              |
++------------------------+--------------+--------------+
+| BSAS                   | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| CLARANS                | $\checkmark$ |              |
++------------------------+--------------+--------------+
+| CLIQUE                 | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| CURE                   | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| DBSCAN                 | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| Elbow                  | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| EMA                    | $\checkmark$ |              |
++------------------------+--------------+--------------+
+| GA (Genetic Algorithm) | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
+| HSyncNet               | $\checkmark$ | $\checkmark$ |
++------------------------+--------------+--------------+
[…]
