### 2019-03-10 #1230 @annoviko
@adavidzh ,
- Paper: yes, it works with `$\checkmark$`, thank you!
- `README`: OPTICS example is updated, visualization example is added.
