### 2019-03-10 annoviko/pyclustering#493 @adavidzh
Great, thanks.
