### 2019-03-10 8e811a6
Clarifications and readability
One things: I have no idea what is meant with "a comprehended interface" in the introduction.
