+++ paper/paper.md
@@ -23,23 +23,23 @@ bibliography: paper.bib
 
 # Introduction
 
-The exponential growth of data leads to an appearance of new disciplines such as cluster analysis, machine learning, and therefore information requires processing in different areas like medicine, commercial, science, engineering, etc. Data processing is performed to extract patterns, clusters, internal structures to understand the nature of the data, make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis. PyClustering mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehended interface that makes it easy to use in every project.
+The exponential growth of data leads to an appearance of new disciplines such as cluster analysis, machine learning, and therefore information requires processing in different areas like medicine, commercial, science, engineering, etc. Data processing is performed to extract patterns, clusters, internal structures to understand the nature of the data, make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis. PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehended interface that makes it easy to use in every project.
 
 # Summary
 
-As it was mentioned in the introduction, PyClustering library is a Python, C++ data mining library focused on cluster analysis. The library provides an implementation of each algorithm and method on Python and C++ programming languages. By default C++ part of the library is used for processing to ensure maximum performance. This is especially relevant for algorithms that are based on oscillatory networks whose dynamic is described by a system of differential equations and where the C++ implementation is more suitable. If PyClustering detects that it is not possible to delegate computation to C++ part, for example, in case of unsupported hardware platform or operating system (current version 0.8.2 supports x86 and x86_64 for Windows and Linux operating systems), then the Python implementation is used. Python supports a wide range of platforms and operating systems and that fact ensures high portability of the library. Such architecture gives a balance between portability and performance. PyClustering uses NumPy [@Oliphant2006] package for computing to increase the performance of the Python implementation. NumPy is the fundamental package for scientific computing that provides efficient operations on large N-dimensional arrays.
+The PyClustering library is a Python and C++ data mining library focused on cluster analysis. The library provides an implementation of each algorithm and method in the Python and C++ programming languages. By default, the C++ part of the library is used for processing in order to achieve maximum performance. This is especially relevant for algorithms that are based on oscillatory networks, whose dynamic is described by a system of differential equations and where the C++ implementation is more suitable. If PyClustering detects that it is not possible to delegate th computation to C++ part, for example, in case of an unsupported hardware platform or operating system (current version 0.8.2 supports x86 and x86_64 for Windows and Linux operating systems), then the Python implementation is used. Python supports a wide range of platforms and operating systems and that ensures high portability of the library. Such an architecture gives a balance between portability and performance. PyClustering uses the NumPy [@Oliphant2006] package to increase the performance of the Python implementation. NumPy is a fundamental package for scientific computing that provides efficient operations on large N-dimensional arrays.
 
[…]
