### 2019-03-10 #1230 @adavidzh
Thanks. I think we're making quick progress.

In https://github.com/annoviko/pyclustering/pull/495 there are several suggestions for the paper.

In the table, I think it would make sense to have a reference for each algorithm implementation.
