### 2019-03-10 annoviko/pyclustering#492 @adavidzh
This is part of the review at https://github.com/openjournals/joss-reviews/issues/1230.
