### 2019-03-11 #1230 @adavidzh
So, for the remaining review points:
1. I see some odd issues in testing, like [in the static code analysis](https://travis-ci.org/annoviko/pyclustering/jobs/504590798#L853) or the [ccore unit tests](https://travis-ci.org/annoviko/pyclustering/jobs/504590799#L2522).
1. I would say that [it is mandatory to demonstrate (via tests)](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) that each algorithm implementations provides the actual expected results. Perhaps this is already part of the unit tests. Let me know
1. Given the Python/C++ options, it would also be relevant to have a performance comparison between the Python and C++ implementations. I think that this could be part of the tests.
