### 2019-03-11 f0da418
#495: Type correction, references are added for each algorithm/method/model, update examples for documentation.
