+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Introduction
 
-The exponential growth of data leads to an appearance of new disciplines such as cluster analysis, machine learning, and therefore information requires processing in different areas like medicine, commercial, science, engineering, etc. Data processing is performed to extract patterns, clusters, internal structures to understand the nature of the data, make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis. PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehended interface that makes it easy to use in every project.
+The exponential growth of data leads to an appearance of new disciplines such as cluster analysis, machine learning, and therefore information requires processing in different areas like medicine, commercial, science, engineering, etc. Data processing is performed to extract patterns, clusters, internal structures to understand the nature of the data, make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis. PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehensive interface that makes it easy to use in every project.
 
 # Summary
 
@@ -41,112 +41,112 @@ The PyClustering library is available on PyPi and from a github repository. Sinc
 
 Algorithms and methods are located in the Python module `pyclustering.cluster` and in the C++ namespace `ccore::clst`.
 
-+------------------------+--------------+--------------+
-| Algorithm              | Python       | C++          |
-+========================+==============+==============+
-| Agglomerative          | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| BANG                   | $\checkmark$ |              |
-+------------------------+--------------+--------------+
-| BIRCH                  | $\checkmark$ |              |
-+------------------------+--------------+--------------+
-| BSAS                   | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| CLARANS                | $\checkmark$ |              |
-+------------------------+--------------+--------------+
-| CLIQUE                 | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| CURE                   | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| DBSCAN                 | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| Elbow                  | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| EMA                    | $\checkmark$ |              |
-+------------------------+--------------+--------------+
-| GA (Genetic Algorithm) | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| HSyncNet               | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| K-Means                | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| K-Means++              | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| K-Medians              | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| K-Medoids (PAM)        | $\checkmark$ | $\checkmark$ |
-+------------------------+--------------+--------------+
-| MBSAS                  | $\checkmark$ | $\checkmark$ |
[…]
