+++ paper/paper.bib
@@ -71,4 +71,510 @@
     publisher  = {ACM},
     address    = {New York, NY, USA},
     keywords   = {Memcheck, Valgrind, dynamic binary analysis, dynamic binary instrumentation, shadow values},
-} 
+}
+
+@book{book::algorithms_for_clustering_data,
+    author          = {Jain, Anil K. and Dubes, Richard C.},
+    title           = {Algorithms for Clustering Data},
+    year            = {1988},
+    isbn            = {0-13-022278-X},
+    publisher       = {Prentice-Hall, Inc.},
+    address         = {Upper Saddle River, NJ, USA},
+}
+
+@inproceedings{inproceedings::bang::1,
+    author          = "Schikuta, Erich and Erhart, Martin",
+    editor          = "Amin, Adnan and Dori, Dov and Pudil, Pavel and Freeman, Herbert",
+    title           = "BANG-Clustering: A novel grid-clustering algorithm for huge data sets",
+    booktitle       = "Advances in Pattern Recognition",
+    year            = "1998",
+    publisher       = "Springer Berlin Heidelberg",
+    address         = "Berlin, Heidelberg",
+    pages           = "867--874",
+    isbn            = "978-3-540-68526-5"
+}
+
+@article{article::birch::1,
+    author          = {Zhang, Tian and Ramakrishnan, Raghu and Livny, Miron},
+    title           = {BIRCH: An Efficient Data Clustering Method for Very Large Databases},
+    journal         = {SIGMOD Rec.},
+    issue_date      = {June 1996},
+    volume          = {25},
+    number          = {2},
+    month           = jun,
+    year            = {1996},
+    issn            = {0163-5808},
+    pages           = {103--114},
+    numpages        = {12},
+    url             = {http://doi.acm.org/10.1145/235968.233324},
+    doi             = {10.1145/235968.233324},
+    acmid           = {233324},
+    publisher       = {ACM},
+    address         = {New York, NY, USA},
+}
+
+@book{book::pattern_recognition::2009,
+    Title           = {Pattern Recognition, fourth edition},
+    Annote          = {SIGNATUR = 2011-10411},
+    Author          = {Sergios Theodoridis and Konstantinos Koutroumbas},
+    Keywords        = {PATTERN RECOGNITION},
+    Publisher       = {Elsevier},
+    Year            = {2009},
+    Aquired         = {2011},
+    Place           = {Favoritenstrasse 9/4th Floor/1863}
+}
+
+@article{article::clarans::1,
+    author          = {Ng, Raymond T. and Han, Jiawei},
+    journal         = {IEEE Transactions on Knowledge and Data Engineering},
+    number          = 5,
+    pages           = {1003--1016},
+    publisher       = {IEEE Computer Society},
+    title           = {CLARANS: A Method for Clustering Objects for Spatial Data Mining},
+    volume          = 14,
+    year            = 2002
+}
+
+@article{article::clique::1,
+    author          = "Agrawal, Rakeshand Gehrke, Johannes and Gunopulos, Dimitrios and Raghavan, Prabhakar",
+    title           = "Automatic Subspace Clustering of High Dimensional Data",
+    journal         = "Data Mining and Knowledge Discovery",
+    year            = "2005",
+    month           = "Jul",
+    day             = "01",
+    volume          = "11",
+    number          = "1",
+    pages           = "5--33",
+    issn            = "1573-756X",
+    doi             = "10.1007/s10618-005-1396-1",
+    url             = "https://doi.org/10.1007/s10618-005-1396-1"
+}
+
+@article{article::cure::1,
+    author          = {Guha, Sudipto and Rastogi, Rajeev and Shim, Kyuseok},
[…]
