+++ docs/citation.bib
@@ -1,13 +1,3 @@
-@book{book::algorithms_for_clustering_data::1988,
-    author          = {Jain, Anil K. and Dubes, Richard C.},
-    title           = {Algorithms for Clustering Data},
-    year            = {1988},
-    isbn            = {0-13-022278-X},
-    publisher       = {Prentice-Hall, Inc.},
-    address         = {Upper Saddle River, NJ, USA},
-}
-
-
 @inproceedings{inproceedings::bang::1,
     author          = "Schikuta, Erich and Erhart, Martin",
     editor          = "Amin, Adnan and Dori, Dov and Pudil, Pavel and Freeman, Herbert",

[…]
