### 2019-03-11 bba6f34
#495: References for each algorithm/method/model in the paper.
