+++ paper/paper.md
@@ -44,55 +44,55 @@ Algorithms and methods are located in the Python module `pyclustering.cluster` a
 +---------------------------------------------------------+--------------+--------------+
 | Algorithm                                               | Python       | C++          |
 +=========================================================+==============+==============+
-| Agglomerative [@book::algorithms_for_clustering_data]   | $\checkmark$ | $\checkmark$ |
+| Agglomerative [@book_algorithms_for_clustering_data]   | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| BANG [@inproceedings::bang::1]                          | $\checkmark$ |              |
+| BANG [@inproceedings_bang_1]                          | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| BIRCH [@article::birch::1]                              | $\checkmark$ |              |
+| BIRCH [@article_birch_1]                              | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| BSAS [@book::pattern_recognition::2009]                 | $\checkmark$ | $\checkmark$ |
+| BSAS [@book_pattern_recognition_2009]                 | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+----------------+------------+
-| CLARANS [@article::clarans::1]                          | $\checkmark$ |              |
+| CLARANS [@article_clarans_1]                          | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| CLIQUE [@article::clique::1]                            | $\checkmark$ | $\checkmark$ |
+| CLIQUE [@article_clique_1]                            | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| CURE [@article::cure::1]                                | $\checkmark$ | $\checkmark$ |
+| CURE [@article_cure_1]                                | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| DBSCAN [@inproceedings::dbscan::1]                      | $\checkmark$ | $\checkmark$ |
+| DBSCAN [@inproceedings_dbscan_1]                      | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| Elbow [@article::cluster::elbow::1]                     | $\checkmark$ | $\checkmark$ |
+| Elbow [@article_cluster_elbow_1]                     | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| EMA [@article::ema::1]                                  | $\checkmark$ |              |
+| EMA [@article_ema_1]                                  | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| GA (Genetic Algorithm) [article::ga::2]                 | $\checkmark$ | $\checkmark$ |
+| GA (Genetic Algorithm) [article_ga_2]                 | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| HSyncNet [@artcile::hsyncnet::1]                        | $\checkmark$ | $\checkmark$ |
+| HSyncNet [@artcile_hsyncnet_1]                        | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| K-Means [@inproceedings::kmeans::1]                     | $\checkmark$ | $\checkmark$ |
+| K-Means [@inproceedings_kmeans_1]                     | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| K-Means++ [@article::kmeans++::1]                       | $\checkmark$ | $\checkmark$ |
+| K-Means++ [@article_kmeans++_1]                       | $\checkmark$ | $\checkmark$ |
[…]
