+++ paper/paper.bib
@@ -28,7 +28,6 @@
     author    = {Oliphant, Travis},
     year      = {2006},
     month     = {01},
-    pages     = {},
     title     = {Guide to NumPy}
 }
 
@@ -73,7 +72,7 @@
     keywords   = {Memcheck, Valgrind, dynamic binary analysis, dynamic binary instrumentation, shadow values},
 }
 
-@book{book::algorithms_for_clustering_data,
+@book{book_algorithms_for_clustering_data,
     author          = {Jain, Anil K. and Dubes, Richard C.},
     title           = {Algorithms for Clustering Data},
     year            = {1988},
@@ -82,7 +81,7 @@
     address         = {Upper Saddle River, NJ, USA},
 }
 
-@inproceedings{inproceedings::bang::1,
+@inproceedings{inproceedings_bang_1,
     author          = "Schikuta, Erich and Erhart, Martin",
     editor          = "Amin, Adnan and Dori, Dov and Pudil, Pavel and Freeman, Herbert",
     title           = "BANG-Clustering: A novel grid-clustering algorithm for huge data sets",
@@ -94,7 +93,7 @@
     isbn            = "978-3-540-68526-5"
 }
 
-@article{article::birch::1,
+@article{article_birch_1,
     author          = {Zhang, Tian and Ramakrishnan, Raghu and Livny, Miron},
     title           = {BIRCH: An Efficient Data Clustering Method for Very Large Databases},
     journal         = {SIGMOD Rec.},
@@ -113,7 +112,7 @@
     address         = {New York, NY, USA},
 }
 
-@book{book::pattern_recognition::2009,
+@book{book_pattern_recognition_2009,
     Title           = {Pattern Recognition, fourth edition},
     Annote          = {SIGNATUR = 2011-10411},
     Author          = {Sergios Theodoridis and Konstantinos Koutroumbas},
@@ -124,7 +123,7 @@
     Place           = {Favoritenstrasse 9/4th Floor/1863}
 }
 
-@article{article::clarans::1,
+@article{article_clarans_1,
     author          = {Ng, Raymond T. and Han, Jiawei},
     journal         = {IEEE Transactions on Knowledge and Data Engineering},
     number          = 5,
@@ -135,7 +134,7 @@
     year            = 2002
 }
 
-@article{article::clique::1,
+@article{article_clique_1,
     author          = "Agrawal, Rakeshand Gehrke, Johannes and Gunopulos, Dimitrios and Raghavan, Prabhakar",
     title           = "Automatic Subspace Clustering of High Dimensional Data",
     journal         = "Data Mining and Knowledge Discovery",
@@ -150,7 +149,7 @@
     url             = "https://doi.org/10.1007/s10618-005-1396-1"
 }
 
-@article{article::cure::1,
+@article{article_cure_1,
     author          = {Guha, Sudipto and Rastogi, Rajeev and Shim, Kyuseok},
     title           = {CURE: An Efficient Clustering Algorithm for Large Databases},
     journal         = {SIGMOD Rec.},
@@ -169,7 +168,7 @@
     address         = {New York, NY, USA},
 }
 
-@inproceedings{inproceedings::dbscan::1,
+@inproceedings{inproceedings_dbscan_1,
     author          = {Ester, Martin and Kriegel, Hans-Peter and Sander, Jorg and Xu, Xiaowei},
     title           = {A Density-based Algorithm for Discovering Clusters a Density-based Algorithm for Discovering Clusters in Large Spatial Databases with Noise},
     booktitle       = {Proceedings of the Second International Conference on Knowledge Discovery and Data Mining},
@@ -183,7 +182,7 @@
     publisher       = {AAAI Press},
 }
 
[…]
