+++ paper/paper.md
@@ -44,55 +44,55 @@ Algorithms and methods are located in the Python module `pyclustering.cluster` a
 +---------------------------------------------------------+--------------+--------------+
 | Algorithm                                               | Python       | C++          |
 +=========================================================+==============+==============+
-| Agglomerative [@book_algorithms_for_clustering_data]   | $\checkmark$ | $\checkmark$ |
+| Agglomerative [@book_algorithms_for_clustering_data]    | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| BANG [@inproceedings_bang_1]                          | $\checkmark$ |              |
+| BANG [@inproceedings_bang_1]                            | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| BIRCH [@article_birch_1]                              | $\checkmark$ |              |
+| BIRCH [@article_birch_1]                                | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| BSAS [@book_pattern_recognition_2009]                 | $\checkmark$ | $\checkmark$ |
-+---------------------------------------------------------+----------------+------------+
-| CLARANS [@article_clarans_1]                          | $\checkmark$ |              |
+| BSAS [@book_pattern_recognition_2009]                   | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| CLIQUE [@article_clique_1]                            | $\checkmark$ | $\checkmark$ |
+| CLARANS [@article_clarans_1]                            | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| CURE [@article_cure_1]                                | $\checkmark$ | $\checkmark$ |
+| CLIQUE [@article_clique_1]                              | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| DBSCAN [@inproceedings_dbscan_1]                      | $\checkmark$ | $\checkmark$ |
+| CURE [@article_cure_1]                                  | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| Elbow [@article_cluster_elbow_1]                     | $\checkmark$ | $\checkmark$ |
+| DBSCAN [@inproceedings_dbscan_1]                        | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| EMA [@article_ema_1]                                  | $\checkmark$ |              |
+| Elbow [@article_cluster_elbow_1]                        | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| GA (Genetic Algorithm) [article_ga_2]                 | $\checkmark$ | $\checkmark$ |
+| EMA [@article_ema_1]                                    | $\checkmark$ |              |
 +---------------------------------------------------------+--------------+--------------+
-| HSyncNet [@artcile_hsyncnet_1]                        | $\checkmark$ | $\checkmark$ |
+| GA - Genetic Algorithm [@article_ga_2]                  | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| K-Means [@inproceedings_kmeans_1]                     | $\checkmark$ | $\checkmark$ |
+| HSyncNet [@artcile_hsyncnet_1]                          | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
-| K-Means++ [@article_kmeans++_1]                       | $\checkmark$ | $\checkmark$ |
+| K-Means [@inproceedings_kmeans_1]                       | $\checkmark$ | $\checkmark$ |
 +---------------------------------------------------------+--------------+--------------+
[…]
