+++ paper/paper.bib
@@ -47,9 +47,8 @@
 @misc{SciPy,
     author    = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
     title     = {{SciPy}: Open source scientific tools for {Python}},
-    year      = {2001--},
-    url       = "http://www.scipy.org/",
-    note      = {[Online; accessed <today>]}
+    year      = {2019},
+    url       = "http://www.scipy.org/"
 }
 
 @article{Nethercote2007,
@@ -279,6 +278,7 @@
 }
 
 @inproceedings{inproceedings_rock_1,
+    author          = {S. {Guha} and R. {Rastogi} and K. {Shim}},
     title           = {ROCK: A Robust Clustering Algorithm for Categorical Attributes},
     booktitle       = {Proceedings of the 15th International Conference on Data Engineering},
     series          = {ICDE '99},
