+++ docs/citation.bib
@@ -215,6 +215,7 @@
 
 
 @inproceedings{inproceedings::rock::1,
+    author          = {S. {Guha} and R. {Rastogi} and K. {Shim}},
     title           = {ROCK: A Robust Clustering Algorithm for Categorical Attributes},
     booktitle       = {Proceedings of the 15th International Conference on Data Engineering},
     series          = {ICDE '99},
