+++ paper/paper.bib
@@ -250,7 +250,7 @@
     pages           = {281--297}
 }
 
-@inproceedings{article_kmeans++_1,
+@inproceedings{article_kmeans_plus_plus_1,
     author          = {David Arthur and Sergei Vassilvitskii},
     title           = {K-means++: the advantages of careful seeding},
     booktitle       = {In Proceedings of the 18th Annual ACM-SIAM Symposium on Discrete Algorithms},
