### 2019-03-11 #1230 @annoviko
@adavidzh, I have fixed typo and added references for each algorithm/method/model that is mentioned in the paper (see post above). About tests: I will describe everything in this evening.
