### 2019-03-11 #1230 @annoviko
Testing is performed using AppVeyor (Windows) and Travis-CI (Linux).

How does it work (Travis-CI - [CI run example](https://travis-ci.org/annoviko/pyclustering/builds/504662001)):
1. **Build CCORE** – C++ part of the library is build for x86 and x86_64 platforms. Warnings and remarks are treated as errors. These binaries are uploaded to yandex cloud. Release candidates are also stored in this cloud.
2. **Run C/C++ static code analyser** – C++ code is verified to find possible mistakes using static analysers, like `scan-build` (from `clang`) and `cppcheck`. Warnings and remarks are considered as errors.
3. **Run unit-tests for CCORE** – C++ part of the library is tested by unit-tests that are written using `GoogleTest`. Each algorithm/method/model is well-tested, there are a lot of samples (see `pyclustering/pyclustering/samples`) that are used to be sure that each algorithm works as it should (functional testing). Almost all unit-tests are functional. Also code coverage is calculated for C++ part and pushed to `coveralls`.
4. **Run memory leak analyser (valgrind) for CCORE** – C++ part of the library is checked to find possible memory leakages. In case of Travis-CI only shock testing is performed, because otherwise it takes more than 25 minutes. Full regression is performed manually before release.
5. **PyClustering unit and integration testing**. Unit-testing is performed for Python code and integration testing for Python + `CCORE` library (C/C++ part). Each algorithm/method/model is tested to be sure that it works as it should using samples from `pyclustering/pyclustering/samples`. Almost all tests are functional. Code coverage is measured and also published to coveralls.
6. **Build documentation** – this job checks that code is documented properly and nothing is forgotten. Each warning, remark is considered as a error.

For Windows (AppVeyor - [CI run example](https://ci.appveyor.com/project/annoviko/pyclustering/builds/22978666)):
1. **BUILD_WINDOWS_CCORE** – build CCORE for x86 and x86_64 and upload binaries to yandex cloud.
2. **UT_WINDOWS_CCORE** – performs unit-testing of `CCORE` as in case of Travis-CI.
3. **BUILD_CYGWIN_CCORE** – build `CCORE` for Cygwin.
4. **UT_CYGWIN_CCORE** – performs unit-testing of `CCORE` for Cygwin (the same scope of tests as in case of job **UT_WINDOWS_CCORE**).
5. **PYCLUSTERING_WINDOWS_X86** – job downloads x86 CCORE from the cloud and performs integration testing.
6.  **PYCLUSTERING_WINDOWS_X64** – job downloads x86_64 CCORE from cloud and performs integration testing.

There are some articles on Wiki related to pyclustering testing:
- [Testing PyClustering](https://github.com/annoviko/pyclustering/wiki/Testing-PyClustering)
- [Unit Testing CCORE](https://github.com/annoviko/pyclustering/wiki/Unit-Testing-CCORE)

I hope this answer covers the first two statements.

About the last statement. There is no such automated tests (to compare performance of Python and C++ code) and I am not sure that they are required. There are some drafts in source code to estimate performance, for example, pyclustering/pyclustering/cluster/examples/dbscan_examples.py – function experiment_execution_time. C++ part of the library is implemented independently from Python using pure C/C++ to provide opportunity to use CCORE’s source code in C/C++ projects that are not related to Python.
