### 2019-03-12 #1230 @adavidzh
Thanks for the details; the post above should probably become part of the documentation.

I think that:
- My question pertains to the Travis steps 3 and 5 above. To assess the testing, I am looking for tests where a fixed outcome (that is known to be good) is compared with the output of the algorithm. Can you please provide links to such tests?
- Regarding the performance, I was thinking exactly along the lines of the [`experiment_execution_time` calls you pointed out](https://github.com/annoviko/pyclustering/blob/619168f4e5b68c21aba7d7d58845543ada5e19f6/pyclustering/cluster/examples/dbscan_examples.py#L218-L219) but having a comprehensive script for all algorithms (not just those with both Python and C++ variants). As concerns this being part of the test suite, it would definitely be useful there, e.g., to check performance changes.
