### 2019-03-12 #1230 @annoviko
**Step 3. Run unit-tests for CCORE.** Tests are located here: `/pyclustering/ccore/tst`, files with prefix `utest-`, for example, `utest-kmeans.cpp`, `utest-rock.cpp` – tests for K-Means and ROCK algorithms correspondingly. These tests can be run using makefile:
```bash
make ut
make utrun
```

**Step 4. Run memory leak analyser (valgrind) for CCORE.** The same scope of tests is used, but compilation is performed using `VALGRIND_ANALYSIS_SHOCK` to reduce amount of tests for this particular job. CI job runs following command:
```bash
make valgrind
```
But full testing is performed using:
```bash
make valgrind_shock
```

Step 5. PyClustering unit and integration testing. Unit-tests are located in following directories (with prefix `ut_`, for example, `ut_bang.py` and `ut_cure.py`):
- `pyclustering/cluster/tests/unit`
- `pyclustering/container/tests/unit`
- `pyclustering/core/tests/`
- `pyclustering/gcolor/tests/`
- `pyclustering/nnet/tests/unit`
- `pyclustering/utils/tests/unit`

Integration tests are located in following directories (with prefix `it_`, for example, `it_dbscan.py` and `it_elbow.py`):
- `pyclustering/cluster/tests/integration`
- `pyclustering/container/tests/integration`
- `pyclustering/nnet/tests/integration`
- `pyclustering/utils/tests/integration`

About performance testing. I can do that, but it takes some time. Should I do this in scope of this review? Or may I do this later?
