### 2019-03-17 #1230 @adavidzh
Thanks. I looked through the tests and they look exactly like what is needed.
Regarding the performance, I don't think it it crucial.

@poulson, let's move to the next steps.
