### 2019-03-22 352161d
#490: Silhouette C++ implementation.
