### 2019-03-22 512fd25
#490: Silhouette K-Searcher prototype.
