### 2019-03-22 aa20ea5
#490: [no-build] Add C++ Silhouette implementation. Added tests that detects bug with K-Means++.
