+++ pyclustering/cluster/examples/kmedians_examples.py
@@ -23,99 +23,99 @@
 
 """
 
-from pyclustering.samples.definitions import SIMPLE_SAMPLES, FCPS_SAMPLES;
+from pyclustering.samples.definitions import SIMPLE_SAMPLES, FCPS_SAMPLES
 
-from pyclustering.cluster import cluster_visualizer;
-from pyclustering.cluster.kmedians import kmedians;
+from pyclustering.cluster import cluster_visualizer
+from pyclustering.cluster.kmedians import kmedians
 
-from pyclustering.utils import draw_clusters;
-from pyclustering.utils import read_sample;
-from pyclustering.utils import timedcall;
+from pyclustering.utils import draw_clusters
+from pyclustering.utils import read_sample
+from pyclustering.utils import timedcall
 
 def template_clustering(start_medians, path, tolerance = 0.25):
-    sample = read_sample(path);
+    sample = read_sample(path)
     
-    kmedians_instance = kmedians(sample, start_medians, tolerance);
-    (ticks, result) = timedcall(kmedians_instance.process);
+    kmedians_instance = kmedians(sample, start_medians, tolerance)
+    (ticks, result) = timedcall(kmedians_instance.process)
     
-    clusters = kmedians_instance.get_clusters();
-    print("Sample: ", path, "\t\tExecution time: ", ticks, "\n");
+    clusters = kmedians_instance.get_clusters()
+    print("Sample: ", path, "\t\tExecution time: ", ticks, "\n")
 
-    draw_clusters(sample, clusters);
+    draw_clusters(sample, clusters)
     
 def cluster_sample1():
-    start_medians = [[3.7, 5.5], [6.7, 7.5]];
-    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE1);
+    start_medians = [[3.7, 5.5], [6.7, 7.5]]
+    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE1)
     
 def cluster_sample2():
-    start_medians = [[3.5, 4.8], [6.9, 7], [7.5, 0.5]];
-    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE2);
+    start_medians = [[3.5, 4.8], [6.9, 7], [7.5, 0.5]]
+    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE2)
     
 def cluster_sample3():
-    start_medians = [[0.2, 0.1], [4.0, 1.0], [2.0, 2.0], [2.3, 3.9]];
-    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE3);
+    start_medians = [[0.2, 0.1], [4.0, 1.0], [2.0, 2.0], [2.3, 3.9]]
+    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE3)
     
 def cluster_sample4():
-    start_medians = [[1.5, 0.0], [1.5, 2.0], [1.5, 4.0], [1.5, 6.0], [1.5, 8.0]];
-    template_clustering(start_medians, SIMPLE_SAMPLES.SAMPLE_SIMPLE4);    
+    start_medians = [[1.5, 0.0], [1.5, 2.0], [1.5, 4.0], [1.5, 6.0], [1.5, 8.0]]
[…]
