### 2019-03-23 0f1a226
#497: Correction for python implementation '[pyclustering.cluster.center_initializer] The same centers or index centers'.
