### 2019-03-24 063e53c
#498: Correction for K-Medians hanging. #497: Correction for same indexes (C++ part).
