### 2019-03-24 1c53d3f
#498: answer_reader - Linux correction.
