### 2019-03-25 ba831b8
#490: Corrections for kmeans++.
