### 2019-03-25 #1230 @adavidzh
> @adavidzh Are you recommending the article for publication?

Yes.
