### 2019-03-25 #1230 @poulson
@annoviko, would you mind handling and/or explaining the five missing DOIs?
