### 2019-03-25 #1230 @poulson
@annoviko Would you mind clarifying that the C++ parallelism is based upon `std::thread` (as evidenced by https://github.com/annoviko/pyclustering/blob/master/ccore/src/parallel/parallel.hpp) and, in addition, perform a thorough editing pass on the Intro and Summary sections? I could point out specific instances if you would prefer, but the issues are somewhat pervasive.
