### 2019-03-25 3e6b203
[paper] Update for references.
