+++ paper/paper.bib
@@ -90,6 +90,7 @@
     address         = "Berlin, Heidelberg",
     pages           = "867--874",
     isbn            = "978-3-540-68526-5"
+    doi             = "10.1007/BFb0033313"
 }
 
 @article{article_birch_1,
@@ -123,14 +124,14 @@
 }
 
 @article{article_clarans_1,
-    author          = {Ng, Raymond T. and Han, Jiawei},
-    journal         = {IEEE Transactions on Knowledge and Data Engineering},
-    number          = 5,
-    pages           = {1003--1016},
-    publisher       = {IEEE Computer Society},
-    title           = {CLARANS: A Method for Clustering Objects for Spatial Data Mining},
-    volume          = 14,
-    year            = 2002
+    author          = {Ng, Raymond and Han, Jiawei},
+    year            = {2002},
+    month           = {10},
+    pages           = {1003- 1016},
+    title           = {CLARANS: A method for clustering objects for spatial data mining},
+    volume          = {14},
+    journal         = {Knowledge and Data Engineering, IEEE Transactions on},
+    doi             = {10.1109/TKDE.2002.1033770}
 }
 
 @article{article_clique_1,
@@ -284,11 +285,13 @@
     series          = {ICDE '99},
     year            = {1999},
     isbn            = {0-7695-0071-4},
-    pages           = {512--},
+    pages           = {512-521},
     url             = {http://dl.acm.org/citation.cfm?id=846218.847264},
     acmid           = {847264},
     publisher       = {IEEE Computer Society},
     address         = {Washington, DC, USA},
+    doi             = {10.1109/ICDE.1999.754967},
+    issn            = {1063-6382},
 }
 
 @article{article_cluster_silhouette_1,
@@ -413,17 +416,12 @@
 }
 
 @article{article_nnet_hysteresis_1,
-    author          = "Jin'no, Kenya",
-    editor          = "In, Visarath and Longhini, Patrick and Palacios, Antonio",
-    title           = "Dynamical Hysteresis Neural Networks for Graph Coloring Problem",
-    bookTitle       = "Applications of Nonlinear Dynamics: Model and Design of Complex Systems",
-    year            = "2009",
-    publisher       = "Springer Berlin Heidelberg",
-    address         = "Berlin, Heidelberg",
-    pages           = "331--340",
-    isbn            = "978-3-540-85632-0",
-    doi             = "10.1007/978-3-540-85632-0_27",
-    url             = "https://doi.org/10.1007/978-3-540-85632-0_27"
+    author          = {Jin'no, Kenya and Taguchi, Hiroshi and Yamamoto, Takao and Hirose, Haruo},
+    year            = {2003},
+    month           = {01},
+    pages           = {737-740},
+    title           = {Dynamical Hysteresis Neural Networks for Graph Coloring Problem},
+    doi             = {10.1109/ISCAS.2003.1206418}
 }
 
 @article{article_legion_1,
