### 2019-03-25 #1230 @annoviko
@poulson, I have added missing DOIs, but I don't know what to do with book 'The design and analysis of spatial data structures'. Looks like there is no DOI for this book, I have checked book itself and https://dl.acm.org/citation.cfm?id=77589, but still the same:
```
@book{Samet:1990:DAS:77589,
 author = {Samet, Hanan},
 title = {The Design and Analysis of Spatial Data Structures},
 year = {1990},
 isbn = {0-201-50255-0},
 publisher = {Addison-Wesley Longman Publishing Co., Inc.},
 address = {Boston, MA, USA},
}
```
I have tried to check references after update and looks like **whedon** has troubles with reference checking - there is no any output.

> ...and, in addition, perform a thorough editing pass on the Intro and Summary sections? I could point out specific instances if you would prefer, but the issues are somewhat pervasive.

Yes, could you please point to them, it would be really useful for me.

> Would you mind clarifying that the C++ parallelism is based upon std::thread

Ok, will add information about parallelism.
