### 2019-03-25 #1230 @adavidzh
> The Design and Analysis of Spatial Data Structures

Perhaps [this](https://doi.org/10.1016/0098-3004(91)90117-V) can help?
