### 2019-03-25 7b98bab
kmedoids does *not* implement the PAM algorithm
c.f. the PAM publication. The other reference is also wrong, it does nowhere seem to mention medoids at all.
