+++ pyclustering/cluster/kmedoids.py
@@ -1,7 +1,7 @@
 """!
 
-@brief Cluster analysis algorithm: K-Medoids (PAM - Partitioning Around Medoids).
-@details Implementation based on papers @cite book::algorithms_for_clustering_data, @cite book::finding_groups_in_data.
+@brief Cluster analysis algorithm: K-Medoids (using a k-means style algorithm).
+@details Implementation of a k-means variation using medoids (but not the original PAM algorithm).
 
 @authors Andrei Novikov (pyclustering@yandex.ru)
 @date 2014-2019
@@ -40,7 +40,7 @@ from pyclustering.core.metric_wrapper import metric_wrapper
 
 class kmedoids:
     """!
-    @brief Class represents clustering algorithm K-Medoids (another one title is PAM - Partitioning Around Medoids).
+    @brief Class represents clustering algorithm K-Medoids (using a k-means style algorithm).
     @details The algorithm is less sensitive to outliers tham K-Means. The principle difference between K-Medoids and K-Medians is that
              K-Medoids uses existed points from input data space as medoids, but median in K-Medians can be unreal object (not from
              input data space).
