### 2019-03-25 816e431
This function does *not* compute the geometric median
Instead, it computed the medoid, the sample with the smallest distance sum.

C.f.: https://en.wikipedia.org/wiki/Geometric_median

The geometric median is usually not restricted to sample points, but continuous in R^d. This is more complex, but allows for better solutions than just the sample points (as in k-means, but requires more computational effort).

The method should be renamed to avoid confusion with the common notion of "median", but as I am editing this in github, I don't have an easy backwards search or refactoring. Therefore, I 
[…]
