### 2019-03-25 30c989c
Update PKG-INFO.rst
Also remove wrong reference to PAM here.
