### 2019-03-25 0647d54
Update README.rst
Also remove wrong reference to PAM here.
