+++ README.rst
@@ -173,7 +173,7 @@ Brief Overview of the Library Content
 +------------------------+---------+-----+
 | K-Medians              | ✓       | ✓   |
 +------------------------+---------+-----+
-| K-Medoids (PAM)        | ✓       | ✓   |
+| K-Medoids (k-means-style) | ✓       | ✓   |
 +------------------------+---------+-----+
 | MBSAS                  | ✓       | ✓   |
 +------------------------+---------+-----+
