### 2019-03-25 9581ad1
Update paper.md
Remove wrong reference.
