### 2019-03-26 #1230 @poulson
Hi @annoviko; I don't mind being more specific at all. Several sections jumped out to me.

For example, there are multiple details worth discussing about the first sentence of the Introduction:

> The exponential growth of data leads to an appearance of new disciplines such as cluster analysis, machine learning, and therefore information requires processing in different areas like medicine, commercial, science, engineering, etc.

First, on a grammatical level:
* "such as cluster analysis, machine learning" should perhaps be "cluster analysis and machine learning".
* A colon would be preferred when listing three or more items, i.e., "medicine, commercial, science, engineering, etc.".

In terms of the message of the sentence: it strikes me as a bit boiler plate without much information content. For example, "The exponential growth of data" begs for an example or explanation for the growth. Such an explanation could be combined with more detailed examples of the different "areas" you are describing: 'commercial' strikes me as a vacuous description.
