### 2019-03-26 #1230 @poulson
For the second sentence:

> Data processing is performed to extract patterns, clusters, internal structures to understand the nature of the data, make decisions or create models that help to predict further behavior.

On a grammatical level:
* "extract patterns, clusters, internal structures" should be "extract patterns, clusters, and internal structures" (with a colon perhaps being preferable).

In terms of the sentence's content:
* 'patterns' and 'internal structures' seems a bit vague and redundant. I would recommend more specific examples.
