### 2019-03-28 1064413
[paper] Update paper in line with remarks from @poulson and @adavidzh.
