+++ paper/paper.md
@@ -23,13 +23,13 @@ bibliography: paper.bib
 
 # Introduction
 
-The exponential growth of data leads to an appearance of new disciplines such as cluster analysis, machine learning, and therefore information requires processing in different areas like medicine, commercial, science, engineering, etc. Data processing is performed to extract patterns, clusters, internal structures to understand the nature of the data, make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis. PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehensive interface that makes it easy to use in every project.
+The exponential growth of data from sensors, devices, mobile networks, social media, and other data sources leads to an appearance of new disciplines such as cluster analysis and machine learning, and therefore information requires processing in different areas like medicine, marketing, science, engineering, etc. Data processing can be performed to extract clusters and structures to understand the nature of the data, for example, finding customer groups with similar needs and identifying their behavior in each group. And using obtained information make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis. PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehensive interface that makes it easy to use in every project.
 
 # Summary
 
 The PyClustering library is a Python and C++ data mining library focused on cluster analysis. The library provides an implementation of each algorithm and method in the Python and C++ programming languages. By default, the C++ part of the library is used for processing in order to achieve maximum performance. This is especially relevant for algorithms that are based on oscillatory networks, whose dynamic is described by a system of differential equations and where the C++ implementation is more suitable. If PyClustering detects that it is not possible to delegate th computation to C++ part, for example, in case of an unsupported hardware platform or operating system (current version 0.8.2 supports x86 and x86_64 for Windows and Linux operating systems), then the Python implementation is used. Python supports a wide range of platforms and operating systems and that ensures high portability of the library. Such an architecture gives a balance between portability and performance. PyClustering uses the NumPy [@Oliphant2006] package to increase the performance of the Python implementation. NumPy is a fundamental package for scientific computing that provides efficient operations on large N-dimensional arrays.
 
-PyClustering provides a separate and highly optimized implementation of clustering algorithms using parallel computing using pure C++, without any third-party code and it can therefore easily integrated in any C++ project as a library or as some part of it. In other words, PyClustering usage is not restricted to the Python language and corresponding dependencies. The C++ implementation of the library is based on the C++14 standard and can be built using common compilers, including gcc, clang, mingw, and VS2015. Such flexibility allows developers or scientists to focus on their own projects and not think about library integration and implementation details.
[…]
