+++ paper/paper.bib
@@ -543,5 +543,6 @@
     year            = {1990},
     isbn            = {0-201-50255-0},
     publisher       = {Addison-Wesley Longman Publishing Co., Inc.},
+    doi             = {https://doi.org/10.1016/0098-3004(91)90117-V}
     address         = {Boston, MA, USA},
 }
