+++ paper/paper.bib
@@ -89,7 +89,7 @@
     publisher       = "Springer Berlin Heidelberg",
     address         = "Berlin, Heidelberg",
     pages           = "867--874",
-    isbn            = "978-3-540-68526-5"
+    isbn            = "978-3-540-68526-5",
     doi             = "10.1007/BFb0033313"
 }
 
