+++ paper/paper.bib
@@ -514,16 +514,6 @@
     address         = {New York, NY, USA},
 }
 
-@article{article_gcolor_hysteresis_1,
-    author          = {Jinnno, Kenya},
-    year            = {2009},
-    month           = {02},
-    pages           = {331-340},
-    title           = {Dynamical Hysteresis Neural Networks for Graph Coloring Problem},
-    volume          = {2009},
-    booktitle       = {Understanding Complex Systems}
-}
-
 @article{article_gcolor_sync_1,
     title           = "Clustering dynamics of nonlinear oscillator network: Application to graph coloring problem",
     journal         = "Physica D: Nonlinear Phenomena",
