### 2019-03-28 ad3805b
#500: Remove PAM from documentation to avoid contradiction.
