### 2019-03-28 #1230 @annoviko
Hello, @poulson

Thank you for your remarks! I have updated the paper in line with them: added examples, rephrased and changed pointed places. Additionally, I have corrected all DOIs (thanks to @adavidzh with DOI for book 'The design and analysis of spatial data structures'). Could you please review updated paper?
