### 2019-04-03 d8b9c22
#496: [pyclustering.cluster.kmedoids]: add stop condition 'max_iterations'.
