+++ ccore/tst/utest-kmeans.cpp
@@ -37,15 +37,17 @@ using namespace ccore::utils::metric;
 
 
 static void
-template_kmeans_length_process_data_common(const dataset_ptr & p_data,
-                                           const dataset & p_start_centers,
-                                           const std::vector<size_t> & p_expected_cluster_length,
-                                           const index_sequence & p_indexes,
-                                           const bool p_observe,
-                                           const distance_metric<point> & p_metric = distance_metric_factory<point>::euclidean_square())
+template_kmeans_length_process_data_common(
+    const dataset_ptr & p_data,
+    const dataset & p_start_centers,
+    const std::vector<size_t> & p_expected_cluster_length,
+    const index_sequence & p_indexes,
+    const bool p_observe,
+    const std::size_t p_itermax = kmeans::DEFAULT_ITERMAX,
+    const distance_metric<point> & p_metric = distance_metric_factory<point>::euclidean_square())
 {
     kmeans_data output_result(p_observe);
-    kmeans solver(p_start_centers, 0.0001, p_metric);
+    kmeans solver(p_start_centers, 0.0001, p_itermax, p_metric);
 
     if (p_indexes.empty()) {
         solver.process(*p_data, output_result);
@@ -58,6 +60,13 @@ template_kmeans_length_process_data_common(const dataset_ptr & p_data,
     const std::size_t dimension = data[0].size();
     const cluster_sequence & actual_clusters = output_result.clusters();
     const dataset & centers = output_result.centers();
+
+    if (p_itermax == 0) {
+        ASSERT_TRUE(actual_clusters.empty());
+        ASSERT_EQ(p_start_centers, centers);
+        ASSERT_EQ(0.0, output_result.wce());
+        return;
+    }
 
     ASSERT_CLUSTER_SIZES(data, actual_clusters, p_expected_cluster_length, p_indexes);
 
@@ -81,42 +90,67 @@ template_kmeans_length_process_data_common(const dataset_ptr & p_data,
 }
 
 
-static void
-template_kmeans_length_process_data_range(const dataset_ptr & p_data,
-                                          const dataset & p_start_centers,
-                                          const std::vector<size_t> & p_expected_cluster_length,
-                                          const index_sequence & p_indexes)
+static void template_kmeans_length_process_data_range(
+    const dataset_ptr & p_data,
+    const dataset & p_start_centers,
+    const std::vector<size_t> & p_expected_cluster_length,
+    const index_sequence & p_indexes)
 {
     template_kmeans_length_process_data_common(p_data, p_start_centers, p_expected_cluster_length, p_indexes, false);
 }
 
 
 static void
-template_kmeans_length_process_data(const dataset_ptr & p_data,
-                                    const dataset & p_start_centers,
-                                    const std::vector<size_t> & p_expected_cluster_length)
+template_kmeans_length_process_data(
+    const dataset_ptr & p_data,
+    const dataset & p_start_centers,
+    const std::vector<size_t> & p_expected_cluster_length)
 {
     template_kmeans_length_process_data_range(p_data, p_start_centers, p_expected_cluster_length, { });
 }
 
 
 static void
-template_kmeans_observer(const dataset_ptr & p_data,
-                         const dataset & p_start_centers,
-                         const index_sequence & p_indexes,
[…]
