### 2019-04-03 3193f34
#474: Refactoring in line with pull request.
