+++ pyclustering/cluster/ga.py
@@ -3,7 +3,7 @@
 @brief Cluster analysis algorithm: Genetic clustering algorithm (GA).
 @details Implementation based on papers @cite article::ga::1, @cite article::ga::2.
 
-@authors Andrey Novikov, Aleksey Kukushkin (pyclustering@yandex.ru)
+@authors Andrei Novikov, Aleksey Kukushkin (pyclustering@yandex.ru)
 @date 2014-2019
 @copyright GNU Public License
 
@@ -58,18 +58,18 @@ class ga_observer:
         """
 
         # Global best chromosome and fitness function for each population
-        self._global_best_result = {'chromosome': [], 'fitness_function': []};
+        self._global_best_result = {'chromosome': [], 'fitness_function': []}
 
         # Best chromosome and fitness function on a population
-        self._best_population_result = {'chromosome': [], 'fitness_function': []};
+        self._best_population_result = {'chromosome': [], 'fitness_function': []}
 
         # Mean fitness function on each population
-        self._mean_ff_result = [];
+        self._mean_ff_result = []
 
         # Flags to collect
-        self._need_global_best = need_global_best;
-        self._need_population_best = need_population_best;
-        self._need_mean_ff = need_mean_ff;
+        self._need_global_best = need_global_best
+        self._need_population_best = need_population_best
+        self._need_mean_ff = need_mean_ff
 
 
     def __len__(self):
@@ -77,11 +77,11 @@ class ga_observer:
         @brief Returns amount of iterations that genetic algorithm was observed.
         
         """
-        global_length = len(self._global_best_result['chromosome']);
-        local_length = len(self._best_population_result['chromosome']);
-        average_length = len(self._mean_ff_result);
+        global_length = len(self._global_best_result['chromosome'])
+        local_length = len(self._best_population_result['chromosome'])
+        average_length = len(self._mean_ff_result)
         
-        return max(global_length, local_length, average_length);
+        return max(global_length, local_length, average_length)
 
 
     def collect_global_best(self, best_chromosome, best_fitness_function):
@@ -94,10 +94,10 @@ class ga_observer:
         """
 
         if not self._need_global_best:
-            return;
+            return
 
-        self._global_best_result['chromosome'].append(best_chromosome);
-        self._global_best_result['fitness_function'].append(best_fitness_function);
+        self._global_best_result['chromosome'].append(best_chromosome)
+        self._global_best_result['fitness_function'].append(best_fitness_function)
 
 
     def collect_population_best(self, best_chromosome, best_fitness_function):
@@ -110,10 +110,10 @@ class ga_observer:
         """
 
         if not self._need_population_best:
-            return;
+            return
 
-        self._best_population_result['chromosome'].append(best_chromosome);
-        self._best_population_result['fitness_function'].append(best_fitness_function);
+        self._best_population_result['chromosome'].append(best_chromosome)
+        self._best_population_result['fitness_function'].append(best_fitness_function)
 
 
     def collect_mean(self, fitness_functions):
@@ -125,9 +125,9 @@ class ga_observer:
         """
 
         if not self._need_mean_ff:
-            return;
+            return
 
-        self._mean_ff_result.append(np.mean(fitness_functions));
+        self._mean_ff_result.append(np.mean(fitness_functions))
 
 
     def get_global_best(self):
@@ -136,7 +136,7 @@ class ga_observer:
                  and its fitness function's value (evolution of global optimum) are stored in lists.
         
         """
-        return self._global_best_result;
+        return self._global_best_result
 
 
     def get_population_best(self):
@@ -145,7 +145,7 @@ class ga_observer:
[…]
