### 2019-04-03 b56d263
#474: Correction for crossover mask generation.
