### 2019-04-06 #1230 @poulson
@annoviko Thank you for having spent the time to incorporate some of the suggestions.

The wording is honestly still quite rough, and, in many places, redundant.

I will provide more detailed comments.

Introduction
=========
You begin with:
"""
The exponential growth of data from sensors, devices, mobile networks, social media, and other data sources leads to an appearance of new disciplines such as cluster analysis and machine learning, and therefore information requires processing in different areas like medicine, marketing, science, engineering etc.
"""
Beyond being a very long sentence involving two separate lists of four items, the statement does not make logical sense to me. In particular, the sentence boils down to the statement "The exponential growth of data in several fields leads to an appearance of new disciplines and therefore information requires processing in different areas". I believe that you could simplify this statement in a form similar to: "A variety of scientific and industrial sectors continue to experience exponential growth in their data volumes, and so automatic categorization techniques have become standard tools for dataset exploration."

Your next sentence is:
"""
Data processing can be performed to extract clusters and structures to understand the nature of the data, for example, finding customer groups with similar needs and identifying their behavior in each group.
"""
This could be rephrased to: "Automatic categorization techniques -- typically referred to as clustering -- help expose the structure of a dataset. For example, the generated clusters might each correspond to a customer group with reasonably similar needs and behavior."

You proceed:
"""
And using obtained information make decisions or create models that help to predict further behavior. As a result, more and more algorithms and methods appear to resolve faced problems.
"""
I would combine these two sentences in a form similar to: "Because the resulting clusters are often used as building blocks for higher-level -- often custom -- predictive models, researchers have continually tweaked and invented new clustering techniques."

Your next sentence is:
"""
PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods including bio-inspired oscillatory networks for data analysis.
"""
I would recommend changing "methods including" to "methods, including" and dropping "for data analysis".

You conclude the paragraph with:
"""
PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users. The library is distributed under GNU Public License and provides a comprehensive interface that makes it easy to use in every project.
"""
Because this first sentence both repeats the previous sentence in beginning with 'PyClustering' and primarily contains the -- in my opinion, implied -- statement that PyClustering is a library for clustering, I would recommend dropping it entirely. I would also recommend dropping the "and provides a comprehensive interface that makes it easy to use in every project": a comprehensive interface is an implied goal of every library, as is ease of use.

Summary
=======
Your first two sentences are:
"""
The PyClustering library is a Python and C++ data mining library focused on cluster analysis. The library provides an implementation of each algorithm and method in the Python and C++ programming language.
"""
I believe the second sentence can be dropped with no significant loss of information.

You proceed:
"""
By default, the C++ part of the library is used for processing in order to achieve maximum performance.
"""
What do you mean by "used for processing"? Do you mean that the expensive computational kernels are written in C++, and that the Python interfaces reuse these C++ kernels?

The next sentence is:
"""
This is especially relevant for algorithms that are based on oscillatory networks, whose dynamic is described by a system of differential equations and where the C++ implementation is more suitable.
"""
I would rephrase the second clause of the sentence to: "whose dynamics are governed by a system of differential equations."

You follow with:
"""
If PyClustering detects that it is not possible to delegate th [sic] computation to C++ part [sic], for example, in case of an unsupported hardware platform or operating system (current version 0.8.2 supports x86 and x86_64 for Windows and Linux operating systems), then the Python implementation is used. Python supports a wide range of platforms and operating systems and that ensures high portability of the library. Such an architecture gives a balance between portability and performance.
"""
I would condense all of these sentences down to just: "If support for a C++ compiler is not detected, PyClustering falls back to pure Python implementations of all kernels."

You conclude the first paragraph of the summary with:
"""
PyClustering uses the NumPy (Oliphant, 2006) package to increase the performance of the Python implementation. NumPy is a fundamental package for scientific computing that provides efficient operations on large N-dimensional arrays.
"""
I do not think that a paper on a python package requires anything more than a citation for NumPy. I would recommend shortening these two sentences to: "In order to increase the performance of the Python implementations, PyClustering makes use of the NumPy (Oliphant, 2006) library for its array manipulations."

Your next paragraph begins:
"""
PyClustering provides a separate and highly optimized implementation of clustering algorithms using parallel computing using C++11, 14 (std::thread, std::async, and in case of Windows functionality that is provided by ppl.h), without any third-party code and it can therefore easily integrated in any C++ project as a library or as some part of it.
"""
Are you implying C++11 or C++14 is required  (this is a minor point anyway, as C++14 is a small extension of C++11)? I would recommend breaking this long sentence into several, more carefully worded, statements. For example: "PyClustering provides optimized, parallel C++14 clustering implementations; on most platforms, threading is provided by std::thread, though the Parallel Patterns Library is used for Windows. Due to the standardization of these threading libraries, PyClustering is simple to integrate into pre-existing projects."

I would entirely delete the subsequent sentences:
"""
In other words, PyClustering usage is not restricted to the Python language and corresponding dependencies. The C++ implementation of the library is based on the C++14 standard and can be built using common compilers, including gcc, clang, mingw, and VS2015. Such flexibility allows developers or scientists to focus on their own projects and not think about library integration and implementation details.
"""

Your next paragraph begins with:
"""
The Python implementation uses the SciPy (Jones, Oliphant, Peterson, & others, 2019), MatPlotLib (Hunter, 2007), NumPy, and Pillow packages. SciPy and NumPy are mandatory dependencies that are used for computing purposes. The MatPlotLib and Pillow packages are optional and are used for visualization services. If these latter two packages are not installed, the PyClustering visualization tools are not available.
"""
I would condense this collection of sentences down to: "The core Python dependencies of PyClustering are NumPy and SciPy (Jones, Oliphant, Peterson, et al., 2019), and MatPlotLib (Hunter, 2007) and Pillow are required for visualization support."

The paragraph closes with:
"""
The PyClustering visualization services display, for example, clustering results, data and its clusters in N-dimensional space, image segments, histograms, algorithm-specific features, dynamics of oscillatory and neural network outputs, etc. Visualization makes the clustering outcome easier to understand and useful for research and educational purposes, especially in case of complex clustering algorithms. For example, in the case of algorithms that are based on oscillatory networks, synchronization processes should be visualized in order to understand the clustering results.
"""
The statement that the visualization services display "clustering results" is unclear to me, as you have not stated *how* the results would be visualized. One could shorten and rephrase these sentences as: "The visualization functionality includes 2D and 3D plots of the cluster embeddings, image segments, and, in the case of oscillatory networks, graphs of the synchronization processes." [With that said, I am unclear on what is meant by 'the synchronization processes".

[To be continued...]
