### 2019-04-06 #1230 @poulson
I would recommend removing the entire next paragraph on oscillatory networks and moving the citations into the earlier references to the technique. I do not feel that the paragraph is elucidating enough of a review of the mechanism behind oscillatory neural networks to warrant inclusion.

For the final paragraph of the Summary: note that 'valgrind' has been misspelled as 'valgri'. With that said: it would make more sense to be explicit with a claim of no errors or warnings being produced by valgrind than simply stating that PyClustering is 'supported' by it. The same holds for cppcheck and scan-build. I would also leave off listing the supported compilers here. Further, the sentence "Code coverage is more than 93% that is ensured by unit and integration tests (there are over 2,200 tests)" could be reworded to "More than 93% code coverage is provided by more than 2200 unit and integration tests." And, lastly, I would drop the final sentence on the API documentation being generated by doxygen.
