### 2019-04-06 48e9024
[paper] Update paper in line with review comments.
