+++ paper/paper.bib
@@ -1,29 +1,3 @@
-@article{Novikov2014,
-    author    = {Novikov, A. V. and Benderskaya, E. N.},
-    title     = {Oscillatory neural networks based on the Kuramoto model for cluster analysis},
-    journal   = {Pattern Recognition and Image Analysis},
-    year      = {2014},
-    month     = {Sep},
-    day       = {01},
-    volume    = {24},
-    number    = {3},
-    pages     = {365--371},
-    abstract  = {This paper presents the results of a study of synchronization processes in oscillatory neural networks of various structures based on the Kuramoto model. The estimates of synchronization processes occurring in the oscillatory networks are examined. The results of studying the practical application of oscillatory networks for solving cluster analysis problems are presented.},
-    issn      = {1555-6212},
-    doi       = {10.1134/S1054661814030146},
-    url       = {https://doi.org/10.1134/S1054661814030146}
-}
-
-@article{Cumin2007,
-    author    = {Unsworth, C.P. and Cumin D.},
-    year      = {2007},
-    month     = {02},
-    title     = {Generalising the Kuramoto Model for the study of Neuronal Synchronisation in the Brain},
-    volume    = {226},
-    journal   = {http://www.esc.auckland.ac.nz/research/tech/esc-tr-638.pdf},
-    doi       = {10.1016/j.physd.2006.12.004}
-}
-
 @book{Oliphant2006,
     author    = {Oliphant, Travis},
     year      = {2006},
