### 2019-04-07 #1230 @annoviko
Hello, @poulson
Thank you very much for detailed remarks related to the paper. I have update the paper in line with your comments.

About your question:

> Do you mean that the expensive computational kernels are written in C++, and that the Python interfaces reuse these C++ kernels?

Each algorithm is implemented using Python and vast majority of algorithms are implemented using C++. If platform is not supported then Python implementation is used for processing, for example, for cluster analysis. If platform is supported, then C++ implementation is called - *.dll or *.so library. So, not just some critical parts are implemented using C++, but whole algorithms.
