+++ pyclustering/cluster/examples/fcm_examples.py
@@ -1,0 +1,209 @@
+"""!
+
+@brief Examples of usage and demonstration of abilities of Fuzzy C-Means algorithm in cluster analysis.
+
+@authors Andrei Novikov (pyclustering@yandex.ru)
+@date 2014-2019
+@copyright GNU Public License
+
+@cond GNU_PUBLIC_LICENSE
+    PyClustering is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    PyClustering is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+@endcond
+
+"""
+
+import random
+
+from pyclustering.samples.definitions import SIMPLE_SAMPLES, FCPS_SAMPLES, FAMOUS_SAMPLES
+
+from pyclustering.cluster import cluster_visualizer_multidim
+from pyclustering.cluster.center_initializer import kmeans_plusplus_initializer
+from pyclustering.cluster.fcm import fcm
+
+from pyclustering.utils import read_sample
+from pyclustering.utils import timedcall
+
+
+def template_clustering(start_centers, path):
+    sample = read_sample(path)
+
+    fcm_instance = fcm(sample, start_centers)
+    fcm_instance.process()
+
+    clusters = fcm_instance.get_clusters()
+    centers = fcm_instance.get_centers()
+    membership = fcm_instance.get_membership()
+
+    print(membership)
+
+    visualizer = cluster_visualizer_multidim()
+    visualizer.append_clusters(clusters, sample)
+    visualizer.append_cluster(centers, marker='*', markersize=10)
+    visualizer.show()
+
+
+def cluster_sample1():
+    start_centers = [[4.7, 5.9], [5.7, 6.5]]
+    template_clustering(start_centers, SIMPLE_SAMPLES.SAMPLE_SIMPLE1)
+
+
+def cluster_sample2():
+    start_centers = [[3.5, 4.8], [6.9, 7], [7.5, 0.5]]
+    template_clustering(start_centers, SIMPLE_SAMPLES.SAMPLE_SIMPLE2)
+
+
+def cluster_sample3():
+    start_centers = [[0.2, 0.1], [4.0, 1.0], [2.0, 2.0], [2.3, 3.9]]
+    template_clustering(start_centers, SIMPLE_SAMPLES.SAMPLE_SIMPLE3)
+
+
+def cluster_sample4():
+    start_centers = [[1.5, 0.0], [1.5, 2.0], [1.5, 4.0], [1.5, 6.0], [1.5, 8.0]]
+    template_clustering(start_centers, SIMPLE_SAMPLES.SAMPLE_SIMPLE4)
+
+
+def cluster_sample5():
+    start_centers = [[0.0, 1.0], [0.0, 0.0], [1.0, 1.0], [1.0, 0.0]]
[…]
