### 2019-04-07 6ad21fe
#386: [pyclustering.cluster.fcm] Temporary substitute matrix calculation of membership to fix the bug.
