+++ pyclustering/cluster/fcm.py
@@ -111,19 +111,15 @@ class fcm:
 
 
     def __update_membership(self):
-        length = len(self.__centers)
-        data_difference = numpy.zeros((length, len(self.__data)))
+        data_difference = numpy.zeros((len(self.__centers), len(self.__data)))
 
         for i in range(len(self.__centers)):
             data_difference[i] = numpy.sum(numpy.square(self.__data - self.__centers[i]), axis=1)
 
         for i in range(len(self.__data)):
-            for j in range(length):
-                divider = numpy.sum([numpy.sum(data_difference[k][i]) for k in range(length)])
-                if data_difference[j][i] == 0.0:
-                    self.__membership[i][j] = 1.0
-                else:
-                    self.__membership[i][j] = 1.0 / (pow(length * data_difference[j][i] / divider, self.__degree))
+            for j in range(len(self.__centers)):
+                divider = sum([pow(data_difference[j][i] / data_difference[k][i], self.__degree) for k in range(len(self.__centers))])
+                self.__membership[i][j] = 1.0 / divider
 
 
     def __calculate_changes(self, updated_centers):
