### 2019-04-08 02f7a86
[no-build] Add paper to 0.8.2.rel branch to create DOI for The Journal of Open Source.
