+++ paper/paper.bib
@@ -1,0 +1,512 @@
+@book{Oliphant2006,
+    author    = {Oliphant, Travis},
+    year      = {2006},
+    month     = {01},
+    title     = {Guide to NumPy}
+}
+
+@article{Hunter2007,
+    author    = {Hunter, J. D.},
+    title     = {Matplotlib: A 2D graphics environment},
+    journal   = {Computing In Science \& Engineering},
+    volume    = {9},
+    number    = {3},
+    pages     = {90--95},
+    abstract  = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.},
+    publisher = {IEEE COMPUTER SOC},
+    doi       = {10.1109/MCSE.2007.55},
+    year      = 2007
+}
+
+@misc{SciPy,
+    author    = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+    title     = {{SciPy}: Open source scientific tools for {Python}},
+    year      = {2019},
+    url       = "http://www.scipy.org/"
+}
+
+@article{Nethercote2007,
+    author     = {Nethercote, Nicholas and Seward, Julian},
+    title      = {Valgrind: A Framework for Heavyweight Dynamic Binary Instrumentation},
+    journal    = {SIGPLAN Not.},
+    issue_date = {June 2007},
+    volume     = {42},
+    number     = {6},
+    month      = jun,
+    year       = {2007},
+    issn       = {0362-1340},
+    pages      = {89--100},
+    numpages   = {12},
+    url        = {http://doi.acm.org/10.1145/1273442.1250746},
+    doi        = {10.1145/1273442.1250746},
+    acmid      = {1250746},
+    publisher  = {ACM},
+    address    = {New York, NY, USA},
+    keywords   = {Memcheck, Valgrind, dynamic binary analysis, dynamic binary instrumentation, shadow values},
+}
+
+@book{book_algorithms_for_clustering_data,
+    author          = {Jain, Anil K. and Dubes, Richard C.},
+    title           = {Algorithms for Clustering Data},
+    year            = {1988},
+    isbn            = {0-13-022278-X},
+    publisher       = {Prentice-Hall, Inc.},
+    address         = {Upper Saddle River, NJ, USA},
+}
+
+@inproceedings{inproceedings_bang_1,
+    author          = "Schikuta, Erich and Erhart, Martin",
+    editor          = "Amin, Adnan and Dori, Dov and Pudil, Pavel and Freeman, Herbert",
+    title           = "BANG-Clustering: A novel grid-clustering algorithm for huge data sets",
+    booktitle       = "Advances in Pattern Recognition",
+    year            = "1998",
+    publisher       = "Springer Berlin Heidelberg",
+    address         = "Berlin, Heidelberg",
+    pages           = "867--874",
+    isbn            = "978-3-540-68526-5",
+    doi             = "10.1007/BFb0033313"
+}
+
+@article{article_birch_1,
+    author          = {Zhang, Tian and Ramakrishnan, Raghu and Livny, Miron},
+    title           = {BIRCH: An Efficient Data Clustering Method for Very Large Databases},
+    journal         = {SIGMOD Rec.},
+    issue_date      = {June 1996},
+    volume          = {25},
+    number          = {2},
+    month           = jun,
+    year            = {1996},
+    issn            = {0163-5808},
+    pages           = {103--114},
+    numpages        = {12},
+    url             = {http://doi.acm.org/10.1145/235968.233324},
+    doi             = {10.1145/235968.233324},
+    acmid           = {233324},
+    publisher       = {ACM},
+    address         = {New York, NY, USA},
+}
+
[…]
