+++ paper/paper.md
@@ -1,0 +1,150 @@
+---
+title: 'PyClustering: Data Mining Library'
+tags:
+  - pyclustering
+  - data mining
+  - clustering
+  - cluster analysis
+  - oscillatory network
+  - machine learning
+  - engineering
+  - python
+  - C++
+authors:
+ - name: Andrei V. Novikov
+   orcid: 0000-0002-9666-9141
+   affiliation: "1"
+affiliations:
+ - name: Independent Researcher
+   index: 1
+date: 22 January 2019
+bibliography: paper.bib
+---
+
+# Introduction
+
+A variety of scientific and industrial sectors continue to experience exponential growth in their data volumes, and so automatic categorization techniques have become standard tools for dataset exploration. Automatic categorization techniques -- typically referred to as clustering -- help expose the structure of a dataset. For example, the generated clusters might each correspond to a customer group with reasonably similar needs and behavior. Because the resulting clusters are often used as building blocks for higher-level -- often custom -- predictive models, researchers have continually tweaked and invented new clustering techniques. PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering algorithms and methods, including bio-inspired oscillatory networks. PyClustering is mostly focused on cluster analysis to make it more accessible and understandable for users.
+
+# Summary
+
+The PyClustering library is a Python and C++ data mining library focused on cluster analysis. By default, the C++ part of the library is used for processing in order to achieve maximum performance. This is especially relevant for algorithms that are based on oscillatory networks, whose dynamics are governed by a system of differential equations. If support for a C++ compiler is not detected, PyClustering falls back to pure Python implementations of all kernels. In order to increase the performance of the Python implementations, PyClustering makes use of the NumPy (Oliphant, 2006) library for its array manipulations.
+
+PyClustering provides optimized, parallel C++14 clustering implementations; on most platforms, threading is provided by std::thread, though the Parallel Patterns Library is used for Windows. Due to the standardization of these threading libraries, PyClustering is simple to integrate into pre-existing projects.
+
+The core Python dependencies of PyClustering are NumPy and SciPy (Jones, Oliphant, Peterson, et al., 2019), and MatPlotLib (Hunter, 2007) and Pillow are required for visualization support. The visualization functionality includes 2D and 3D plots of the cluster embeddings, image segments, and, in the case of oscillatory networks, graphs of the synchronization processes.
+
+The PyClustering library is available on PyPi and from a github repository. Since the first release on PyPi in 2014, it has been downloaded more than 141.000 times. The quality of the library is supported by static and dynamic analyzers, such as cppcheck, scan-build, and valgrind [@Nethercote2007]. More than 93% code coverage is provided by more than 2200 unit and integration tests. Each commit to the repository triggers building, analysis, and testing on CI services such as travis-ci or appveyor. PyClustering provides fully-documented code for each library version, including examples, math and algorithms description, and installation instructions.
+
+# Clustering Algorithms
+
+Algorithms and methods are located in the Python module `pyclustering.cluster` and in the C++ namespace `ccore::clst`.
+
++---------------------------------------------------------+--------------+--------------+
+| Algorithm                                               | Python       | C++          |
++=========================================================+==============+==============+
+| Agglomerative [@book_algorithms_for_clustering_data]    | $\checkmark$ | $\checkmark$ |
++---------------------------------------------------------+--------------+--------------+
+| BANG [@inproceedings_bang_1]                            | $\checkmark$ |              |
++---------------------------------------------------------+--------------+--------------+
+| BIRCH [@article_birch_1]                                | $\checkmark$ |              |
++---------------------------------------------------------+--------------+--------------+
+| BSAS [@book_pattern_recognition_2009]                   | $\checkmark$ | $\checkmark$ |
++---------------------------------------------------------+--------------+--------------+
[…]
