### 2019-04-08 #1230 @annoviko
Hello, Jack (@poulson) ,

There is already DOI for 0.8.2 release that is described in the paper: https://zenodo.org/record/1491324#.XKszIlUzaUk

```
    doi          = {10.5281/zenodo.1491324},
    url          = {https://doi.org/10.5281/zenodo.1491324}
```
