### 2019-04-09 ee79fe4
#386: Documentation for python code.
