+++ setup.py
@@ -80,7 +80,7 @@ setup(
       author_email='pyclustering@yandex.ru',
 
       install_requires=['scipy', 'matplotlib', 'numpy', 'Pillow'],
-      python_requires='>=3.4',
+      python_requires='>=3.5',
       package_data={
                       'pyclustering.samples': ['samples/famous/*.data', 'samples/famous/*.answer', 'samples/fcps/*.data', 'samples/fcps/*.answer', 'samples/simple/*.data', 'samples/simple/*.answer', 'graphs/*.grpr', 'images/*.png', 'images/digits/*.png'],
                       'pyclustering.core': ['x64/linux/ccore.so', 'x86/linux/ccore.so',
