### 2019-04-10 #1230 @labarba
The Zenodo archive in DOI 10.5281/zenodo.1491324 is Version 0.8.2 from Nov 19, 2018. If you've made changes during this review, then you would have to make a new tagged version in your software repository, and an update to the Zenodo deposit.
