### 2019-04-10 efaf705
#386: Fuzzy C-Means algorithm (C++ implementation).
