### 2019-04-11 8170b58
#386: Interface C++ <-> Python. Unit-tests for C++ implementation. Corrections.
