+++ ccore/tst/utest-fcm.cpp
@@ -1,0 +1,239 @@
+/**
+*
+* @authors Andrei Novikov (pyclustering@yandex.ru)
+* @date 2014-2019
+* @copyright GNU Public License
+*
+* GNU_PUBLIC_LICENSE
+*   pyclustering is free software: you can redistribute it and/or modify
+*   it under the terms of the GNU General Public License as published by
+*   the Free Software Foundation, either version 3 of the License, or
+*   (at your option) any later version.
+*
+*   pyclustering is distributed in the hope that it will be useful,
+*   but WITHOUT ANY WARRANTY; without even the implied warranty of
+*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+*   GNU General Public License for more details.
+*
+*   You should have received a copy of the GNU General Public License
+*   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*
+*/
+
+
+#include "gtest/gtest.h"
+
+#include "samples.hpp"
+
+#include "cluster/fcm.hpp"
+
+#include "utenv_check.hpp"
+
+#include <cmath>
+#include <numeric>
+
+
+using namespace ccore::clst;
+
+
+static void template_fcm_data_processing(
+    const dataset_ptr & p_data,
+    const dataset & p_start_centers,
+    const double p_m,
+    const std::vector<size_t> & p_expected_cluster_length,
+    const std::size_t p_itermax = fcm::DEFAULT_ITERMAX)
+{
+    fcm_data output_result;
+    fcm solver(p_start_centers, p_m, fcm::DEFAULT_TOLERANCE, p_itermax);
+
+    solver.process(*p_data, output_result);
+
+    if (p_itermax == 0) {
+        ASSERT_TRUE(output_result.clusters().empty());
+        ASSERT_TRUE(output_result.membership().empty());
+        ASSERT_EQ(p_start_centers, output_result.centers());
+        return;
+    }
+
+    ASSERT_CLUSTER_SIZES(*p_data, output_result.clusters(), p_expected_cluster_length);
+
+    const std::size_t dimension = p_data->at(0).size();
+    for (const auto & center : output_result.centers()) {
+        ASSERT_EQ(dimension, center.size());
+
+        for (std::size_t dim = 0; dim < center.size(); dim++) {
+            ASSERT_FALSE(std::isnan(center[dim]));
+        }
+    }
+
+    ASSERT_EQ(output_result.centers().size(), output_result.clusters().size());
+
+    for (const auto & probablities : output_result.membership()) {
+        double total_probability = std::accumulate(probablities.begin(), probablities.end(), 0.0);
+        ASSERT_DOUBLE_EQ(1.0, total_probability);
+    }
+}
+
+
+TEST(utest_fcm, allocation_sample_simple_01) {
+    dataset start_centers = { { 3.7, 5.5 },{ 6.7, 7.5 } };
+    std::vector<size_t> expected_clusters_length = { 5, 5 };
+    template_fcm_data_processing(simple_sample_factory::create_sample(SAMPLE_SIMPLE::SAMPLE_SIMPLE_01), start_centers, 2, expected_clusters_length);
+}
+
+
+TEST(utest_fcm, allocation_sample_simple_01_centers_are_points) {
[…]
