### 2019-04-11 dfc57ee
#494: Update warnings for 0.8.2.rel.
