### 2019-04-11 #1230 @annoviko
Hello, @labarba, there were changes related to project description, functionality wasn't changed. But anyway, I have prepare special 0.8.2 JOSS release with these changes if it is required. Here is DOI: **10.5281/zenodo.2636831** - version 0.8.2.
