### 2019-04-12 a7e2406
#386: Unit, integration tests for FCM. Corrections.
