+++ pyclustering/__init__.py
@@ -1,6 +1,7 @@
 """!
 
-@brief PyClustering module that consists of general modules related to clustering, graph coloring, containers, neural networks, oscillatory networks.
+@brief PyClustering module that consists of general modules related to clustering, graph coloring, containers, neural
+        networks, oscillatory networks.
 
 @authors Andrei Novikov (pyclustering@yandex.ru)
 @date 2014-2019
@@ -24,12 +25,17 @@
 @mainpage PyClustering library
 
 @section intro_sec Introduction
-PyClustering library is a collection of cluster analysis, graph coloring, travelling salesman problem algorithms, oscillatory and neural network models, containers, 
-tools for visualization and result analysis, etc. High performance is ensured by CCORE library that is a part of the pyclustering library where almost the same algorithms,
-models, tools are implemented. There is ability to use python code implementation only or CCORE (C/C++) implementation using special flag. CCORE library does not
-use python.h interface to communicate with python code due to requirement to save ability to use CCORE library or C/C++ code part of CCORE in other projects.
+PyClustering is an open source data mining library written in Python and C++ that provides a wide range of clustering
+algorithms and methods, including bio-inspired oscillatory networks. PyClustering is mostly focused on cluster analysis
+to make it more accessible and understandable for users. The library is distributed under GNU Public License and
+provides a comprehensive interface that makes it easy to use in every project.
 
-PyClustering consists of five general modules where the algorithms, models, tools are placed:
+By default, the C++ part of the library is used for processing in order to achieve maximum performance. This is
+especially relevant for algorithms that are based on oscillatory networks, whose dynamics are governed by a system of
+differential equations. If support for a C++ compiler is not detected, PyClustering falls back to pure Python
+implementations of all kernels.
+
+PyClustering consists of five general modules.
 
 Cluster analysis algorithms and methods (module pyclustering.cluster):
 - Agglomerative (pyclustering.cluster.agglomerative);
@@ -42,6 +48,7 @@ Cluster analysis algorithms and methods (module pyclustering.cluster):
 - DBSCAN (pyclustering.cluster.dbscan);
 - Elbow (pyclustering.cluster.elbow);
 - EMA (pyclustering.cluster.ema);
+- Fuzzy C-Means (pyclustering.cluster.fcm);
 - GA (genetic algorithm pyclustering.cluster.ga);
 - HSyncNet (bio-inspired algorithm pyclustering.cluster.hsyncnet);
 - K-Means (pyclustering.cluster.kmeans);
@@ -79,7 +86,7 @@ Containers (module pyclustering.container):
 - KD-Tree (pyclustering.container.kdtree);
 
 
-Utils that can be used for analysis, visualization, etc are placed in module pyclustering.utils.
+Utils (pyclustering.utils) that can be used for analysis, visualization, etc.
 
 
 @section install_sec Installation
@@ -93,7 +100,8 @@ The library can be compiled and manually installed on linux machine wherever you
     # extract content of the pyclustering library...
     # compile CCORE library (core of the pyclustering library).
     cd pyclustering/ccore
-    make ccore
+    make ccore_x64 # if platform is x86_64 (64-bit)
+    # make ccore_x86 # if platform is x86 (32-bit)
     
     # return to parent folder of the pyclustering library
     cd ../
@@ -103,91 +111,90 @@ The library can be compiled and manually installed on linux machine wherever you
     export PYTHONPATH=${PYTHONPATH}
 @endcode
 
[…]
