### 2019-04-12 1c788ec
#386: Warnings and remarks correction.
