### 2019-04-13 e6cac2f
#486: MAC OS support (MAC OS 10.12 Sierra 64-bit is used for testing).
