### 2019-04-13 43c6149
#487: [documentation] Detailed description of user-defined metric.
