+++ .travis.yml
@@ -1,5 +1,4 @@
 dist: trusty
-language: python
 
 
 os:
