+++ ci/travis-ci.sh
@@ -262,6 +262,10 @@ run_build_test_ccore_macos_job() {
     # build ccore library
     build_ccore x64
 
+    # install python and corresponding packages
+    brew install python3
+    pip3 install numpy matplotlib scipy Pillow
+
     # run integration tests
     python pyclustering/tests/tests_runner.py --integration
 
