### 2019-04-13 b16eb06
#486: MacOS support description.
