+++ setup.py
@@ -38,7 +38,7 @@ def load_readme():
 setup(
       name='pyclustering',
       packages=find_packages(),
-      version='0.8.2',
+      version='0.9.0',
       description='pyclustring is a python data mining library',
       long_description=load_readme(),
       url='https://github.com/annoviko/pyclustering',
@@ -57,17 +57,13 @@ setup(
                      'Intended Audience :: Science/Research',
                      'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
                      'Natural Language :: English',
-                     'Operating System :: Microsoft :: Windows :: Windows 7',
-                     'Operating System :: Microsoft :: Windows :: Windows 8',
-                     'Operating System :: Microsoft :: Windows :: Windows 10',
+                     'Operating System :: Microsoft',
                      'Operating System :: POSIX :: Linux',
                      'Operating System :: Unix',
+                     'Operating System :: MacOS',
                      'Programming Language :: C',
                      'Programming Language :: C++',
-                     'Programming Language :: Python :: 3.4',
-                     'Programming Language :: Python :: 3.5',
-                     'Programming Language :: Python :: 3.6',
-                     'Programming Language :: Python :: 3.7',
+                     'Programming Language :: Python :: 3',
                      'Topic :: Education',
                      'Topic :: Scientific/Engineering :: Artificial Intelligence',
                      'Topic :: Scientific/Engineering :: Bio-Informatics',
@@ -84,7 +80,8 @@ setup(
       package_data={
                       'pyclustering.samples': ['samples/famous/*.data', 'samples/famous/*.answer', 'samples/fcps/*.data', 'samples/fcps/*.answer', 'samples/simple/*.data', 'samples/simple/*.answer', 'graphs/*.grpr', 'images/*.png', 'images/digits/*.png'],
                       'pyclustering.core': ['x64/linux/ccore.so', 'x86/linux/ccore.so',
-                                            'x64/win/ccore.dll', 'x86/win/ccore.dll'],
+                                            'x64/win/ccore.dll', 'x86/win/ccore.dll',
+                                            'x64/macos/ccore.so'],
                    },
 
       data_files=[('', ['LICENSE', 'CHANGES', 'README.rst', 'PKG-INFO.rst'])],
