### 2019-04-14 #1230 @labarba
Congratulations, @annoviko, your JOSS paper is now published!

Big thanks to our editor: @poulson, and reviewer: @adavidzh — your valuable contribution makes JOSS possible! 🙏
